% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{apply_cat_palette_legend}
\alias{apply_cat_palette_legend}
\title{Create an svg simple legend for a categorical palette}
\usage{
apply_cat_palette_legend(palette, data)
}
\arguments{
\item{palette}{}

\item{data}{}
}
\description{
Colors can be specified as palettes or as colors, therefore we need to cover both.
}
\details{
TODO: Consider not passing palettes and only passing colors inside the data
}
\keyword{internal}
