% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{boxplot_chart}
\alias{boxplot_chart}
\title{ggplot for a set of faceted boxplots}
\usage{
boxplot_chart(ds, violin, show_points, log_project_y, title_data = NULL)
}
\arguments{
\item{ds}{\code{data.frame()}

A dataframe as output by \link{bp_subset_data}}

\item{violin}{\code{logical(1)}

Shows a violin plot instead of a boxplot}

\item{show_points}{\code{logical(1)}

Shows individual points in the boxplot chart}

\item{title_data}{\code{list()}

Shows a title summarising parameter values.}
}
\value{
A ggplot chart
}
\description{
The chart consists of a faceted set of boxplots for biomarker data
}
\details{
\subsection{Labels:}{
\itemize{
\item X axis and color aesthetic are labelled using the label atttribute from \code{main_group} column from \code{ds}
\item Y axis is labelled using the label atttribute from \code{value} column from \code{ds}
}
}

\subsection{Aesthetics:}{
\itemize{
\item \code{parameter} is coded as 1st level row facets
\item \code{main_group} is coded as colors and x axis
\item \code{sub_group} is coded as 1st level column facets
\item \code{page_group} is coded as 2nd level row facets
}
}

\subsection{Hack:}{

\link[shiny:brushedPoints]{shiny::nearPoints} do not manage well charts with no x aesthetic. Therefore when no \code{main_group} is present
in \code{ds} a dummy one is added.
}
}
\keyword{internal}
