% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{boxplot_composed}
\alias{boxplot_composed}
\alias{bp_get_boxplot_output}
\alias{bp_get_listings_output}
\alias{bp_get_single_listings_output}
\alias{bp_get_count_output}
\alias{bp_get_summary_output}
\alias{bp_get_significance_output}
\title{Composes data selection and charting for boxplot}
\usage{
bp_get_boxplot_output(ds, violin, show_points, log_project_y, title_data)

bp_get_listings_output(ds, closest_point)

bp_get_single_listings_output(ds, closest_point, input_id)

bp_get_count_output(ds)

bp_get_summary_output(ds)

bp_get_significance_output(ds)
}
\arguments{
\item{ds}{\code{data.frame()}

A dataframe as output by \link{bp_subset_data}}

\item{violin}{\code{logical(1)}

Shows a violin plot instead of a boxplot}

\item{show_points}{\code{logical(1)}

Shows individual points in the boxplot chart}

\item{title_data}{\code{list()}

Shows a title summarising parameter values.}

\item{input_id}{Shiny ID for the single listing button}
}
\description{
This is a set of \emph{glue} functions that in most cases follow the pattern
of receiving data, calculating data in an independent/several independent functions
}
\details{
In some cases aesthetic or simple error validations can occur

This set of functions is not directly tested as they are indirectly tested in the running app
}
\keyword{internal}
