% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{bp_count_table}
\alias{bp_count_table}
\title{Counts the number of rows grouped by all variables except \code{subject_id} and \code{value}}
\usage{
bp_count_table(ds)
}
\arguments{
\item{ds}{\code{data.frame()}
A data frame to count the rows over.}
}
\value{
\code{data.frame()}
A data frame with the counts for each group.
}
\description{
\itemize{
\item Counts the number of rows grouped by all variables except \code{subject_id} and \code{value}.
\item The function returns a data frame with the counts for each group.
}
}
\keyword{boxplot,}
\keyword{internal}
