% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{bp_get_closest_gen_click}
\alias{bp_get_closest_gen_click}
\alias{bp_get_closest_single_click}
\alias{bp_get_closest_double_click}
\title{Clicking helpers}
\usage{
bp_get_closest_gen_click(ds, click)

bp_get_closest_single_click(ds, click)

bp_get_closest_double_click(ds, click)
}
\arguments{
\item{ds}{\code{data.frame()}

A data frame containing the data to be filtered.}

\item{click}{\code{list()}

A list containing the click information as received from the \code{Shiny} input}
}
\value{
\code{data.frame}

A data frame containing the closest row in \code{ds} to the click.
}
\description{
The boxplot requires several helpers to manage the clicks. The main function is \code{bp_get_closest_gen_click}, and the
other two are specific calls for \code{click} and \code{double_click} events.
\subsection{bp_get_closest_gen_click}{

This function takes a dataset and a general click and returns the closest row to the click. It uses the
\link[shiny:brushedPoints]{shiny::nearPoints}. \link[shiny:brushedPoints]{shiny::nearpoints} function is not particularly well-behaved in two cases:
\itemize{
\item The ggplot does contain several facet levels
Solved by:
Prefiltering \code{ds} manually so it only contains the rows corresponding to the clicked facet
\item The ggplot lacks an \code{x} or \code{y} aesthetic
Solved by:
In this particular case the \code{x} aesthetic correspond to the \code{CNT$MAIN_GROUP} column. If no grouping
is specified during the selection this function assumes that a dummy \code{x} aesthetic was used in \link{boxplot_chart}.
Therefore, this dummy variable is introduced as a column in \code{ds}, and \link[shiny:brushedPoints]{shiny::nearPoints} can be used.
\item The aesthetic has been defined using the \code{.data} pronoun to overcome NSE used in \link[ggplot2:ggplot]{ggplot2::ggplot}
Solved by:
Stripping \code{.data} pronouns before passing them into \link[shiny:brushedPoints]{shiny::nearPoints}
}
}

\subsection{Specific calls:}{

After a click is made and a table shown, when the dataset changes, the app reaches this function in an invalid state
where \code{ds} contains the new dataset and \code{click} contains the click from the previous dataset. This usually provokes
that a 0-row dataset is returned.
}
}
\keyword{internal}
