% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{bp_listings_table}
\alias{bp_listings_table}
\title{Subsets a data.frame based on the values of a one-rowed data.frame}
\usage{
bp_listings_table(ds, f_ds)
}
\arguments{
\item{ds}{\code{data.frame()}

data.frame to be subset}

\item{f_ds}{\code{data.frame()}

one-rowed data.frame}
}
\value{
\code{data.frame()}

The subset dataset
}
\description{
It subsets \code{ds} based on the values of \code{f_ds}.
}
\details{
No check no names of data.frames is performed (e.g.: names in \code{f_ds} are names in \code{ds})
}
\keyword{boxplot,}
\keyword{internal}
