% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{bp_significance_table}
\alias{bp_significance_table}
\title{Create a table of t-test results for pairwise comparisons for \code{main_group} in a data frame}
\usage{
bp_significance_table(ds)
}
\arguments{
\item{ds}{\code{data.frame()}
A data frame to be analyzed.}
}
\value{
\code{data.frame()}
A data frame with the test results.
The returned data frame will have one row for each pairwise comparison and group, and three additional columns
\code{Test}, \code{Comparison}, and \verb{P Value}.
}
\description{
Creates a table of t-test results for pairwise comparisons for the groups defined by \code{main_group}
groups in a data frame and values in \code{value}

Each of the pairwise comparison is performed for each of groups defined by the columns that are not
\code{main_group}, \code{value} and \code{subject_id}.

If no or a single data point for a group is present for a given group \code{NA} values are returned
}
\keyword{internal}
