% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{bp_summary_table}
\alias{bp_summary_table}
\title{Calculates a set of summary statistics grouped by all variables except \code{subject_id} and \code{value}}
\usage{
bp_summary_table(ds)
}
\arguments{
\item{ds}{\code{data.frame()}
A data frame to perform the summary over.}
}
\value{
\code{data.frame()}
A data frame with the summary statistics.
}
\description{
\itemize{
\item Calculates a set of summary statistics grouped by all variables except \code{subject_id} and \code{value}.
\item The summary statistics include \code{N}, \code{Mean}, \code{Median}, \code{SD}, \code{Min}, \code{Q1}, \code{Median}, \code{Q3}, \code{Max} and \verb{NA Values}.
\item \code{NA} values are ignored when calculating stats
}
}
\keyword{internal}
