% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_corr_hm.R
\name{ch_subset_data}
\alias{ch_subset_data}
\title{Subset datasets for correlation heatmap}
\usage{
ch_subset_data(
  sel,
  cat_col,
  par_col,
  val_col,
  vis_col,
  bm_ds,
  subj_col,
  anlfl_col = NULL
)
}
\arguments{
\item{sel}{a data.frame with three columns CNT$CAT, CMT$PAR, CNT$VIS.}

\item{val_col}{\verb{[character*ish*(1)]}

Column containing values for the parameters}

\item{bm_ds}{\code{data.frame}

data frames to be used as inputs in \link{subset_bds_param}}

\item{subj_col, par_col, cat_col, vis_col, anlfl_col}{\verb{[character(1)]}

Column for subject id, category, parameter, visit and analysis flag. All
specified columns must be factors. Analysis flag is optional.}
}
\value{
\verb{[data.frame()]}\tabular{lllll}{
   \code{category} \tab \code{parameter} \tab \code{subject_id} \tab \code{value} \tab \code{visit} \cr
   xx \tab xx \tab xx \tab xx \tab xx \cr
}
}
\description{
Prepares the main data frame  for the rest of the correlation heatmap functions by
subsetting \code{bm_dataset} according to the category, parameter and visit combinations
captured by the rows of the \code{sel} parameter.
}
\details{
\itemize{
\item factors from \code{bm_ds} are releveled so all extra levels not present after subsetting are dropped and are sorted
according to \code{par} and \code{cat}. Unless parameters are renamed in \code{\link[=subset_bds_param]{subset_bds_param()}} then no releveling occurs.
\item \code{label} attributes from \code{bm_ds} are retained when available.
}
}
\keyword{internal}
