% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{check_selected_choices}
\alias{check_selected_choices}
\title{If selected is not contained in choices send a warning}
\usage{
check_selected_choices(selected, choices, inputId)
}
\arguments{
\item{selected}{selection set}

\item{choices}{choices set}

\item{inputId}{the inputId of the shiny element being checked}
}
\description{
If selected is not contained in choices send a warning
}
\keyword{internal}
