% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{composed}
\alias{composed}
\alias{get_roc_plot_output}
\alias{get_dens_plot_output}
\alias{get_histo_plot_output}
\alias{get_raincloud_output}
\alias{get_metrics_output}
\alias{get_gt_summary_output}
\alias{get_info_panel_output}
\title{Composed functions}
\usage{
get_roc_plot_output(ds_list, param_as_cols, fig_size, is_sorted)

get_dens_plot_output(ds, param_as_cols, fig_size)

get_histo_plot_output(ds, param_as_cols, fig_size)

get_raincloud_output(ds, param_as_cols, fig_size)

get_metrics_output(
  ds,
  param_as_cols,
  fig_size,
  x_metrics_col,
  compute_metric_fn
)

get_gt_summary_output(ds_list, sort_alph)

get_info_panel_output(ds)
}
\description{
Functions that creates and output to be displayed in the app, usually, by composing data and plotting calls.
}
\section{Functions}{
\itemize{
\item \code{get_roc_plot_output()}: ROC plot

\item \code{get_dens_plot_output()}: Density plot

\item \code{get_histo_plot_output()}: Density plot

\item \code{get_raincloud_output()}: Raincloud plot

\item \code{get_metrics_output()}: Metrics plot

\item \code{get_gt_summary_output()}: GT Summary

\item \code{get_info_panel_output()}: Info Panel

}}
\keyword{composed}
\keyword{internal}
