% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{compute_metric}
\alias{compute_metric}
\alias{assert_compute_metric_data}
\alias{compute_metric_data}
\title{Helpers for computing metric data from the subset dataset}
\usage{
assert_compute_metric_data(r)

compute_metric_data(predictor, response)
}
\arguments{
\item{r}{\verb{[data.frame()]}

dataframe resulting from compute_metric_data}

\item{predictor}{\verb{[numeric()]}

The scores of the predictor}

\item{response}{\verb{[factor()]}

The response value}
}
\value{
\verb{[data.frame()]}

With columns:
\itemize{
\item \code{type} \verb{[character()]}: Metric name
\item \code{y} \verb{[numeric()]}: Metric value
\item \code{score}, \code{norm_score}, \code{norm_rank} \verb{[numeric()]}: Raw, normalized, and normalized rank predictor value per group.
}

With an attribute:
\itemize{
\item \code{limits} \verb{[list()]}: With one entry per metric type. Each entry is a \code{numeric(2)} that contains the plotting limits for the metric.
}
}
\description{
Helpers for computing metric data from the subset dataset
}
\section{Functions}{
\itemize{
\item \code{assert_compute_metric_data()}: Assert compute_metric result data.frame types are correct

\item \code{compute_metric_data()}: Compute metric analysis

}}
