% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_corr_hm.R
\name{corr_hm_server}
\alias{corr_hm_server}
\title{Correlation heatmap server function}
\usage{
corr_hm_server(
  id,
  bm_dataset,
  subjid_var = "USUBJID",
  cat_var = "PARCAT",
  par_var = "PARAM",
  visit_var = "AVISIT",
  value_vars = "AVAL",
  default_value = NULL
)
}
\arguments{
\item{id}{\verb{[character(1)]}

Shiny ID}

\item{bm_dataset}{\verb{[data.frame()]}

An ADBM-like dataset similar in structure to the one in
\href{https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192}{this example},
with one record per subject per parameter per analysis visit.

It should have, at least, the columns specified by the parameters \code{subjid_var}, \code{cat_var},
\code{par_var}, \code{visit_var} and \code{value_vars}.
The semantics of these columns are as described in the CDISC standard for variables
USUBJID, PARCAT, PARAM, AVISIT and AVAL, respectively.}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to the subject ID}

\item{cat_var, par_var, visit_var}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{default_value}{\verb{[character(1)|NULL]}

Default values for the selectors}
}
\description{
Correlation heatmap server function
}
\keyword{developers}
