% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_corr_hm.R
\name{corr_hm_UI}
\alias{corr_hm_UI}
\title{Correlation heatmap UI function}
\usage{
corr_hm_UI(
  id,
  default_cat = NULL,
  default_par = NULL,
  default_visit = NULL,
  default_corr_method = NULL
)
}
\arguments{
\item{id}{\verb{[character(1)]}

Shiny ID}
}
\description{
Correlation heatmap UI function
}
\keyword{developers}
