% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineplot.R
\docType{data}
\name{default_lineplot_fns}
\alias{default_lineplot_fns}
\alias{lp_mean_summary_fns}
\alias{lp_median_summary_fns}
\title{Default lineplot summary functions}
\format{
An object of class \code{list} of length 3.

An object of class \code{list} of length 3.
}
\usage{
lp_mean_summary_fns

lp_median_summary_fns
}
\description{
Default lineplot summary functions
}
\section{Functions}{
\itemize{
\item \code{lp_mean_summary_fns}: Default mean functions

\item \code{lp_median_summary_fns}: Default median functions

}}
\keyword{datasets}
