% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_logic.R
\name{equal_and_mask_from_vec}
\alias{equal_and_mask_from_vec}
\title{Create a mask to filter a dataframe from a named list of values}
\usage{
equal_and_mask_from_vec(ds, fl)
}
\arguments{
\item{ds}{\code{data.frame()}

A dataframe}

\item{fl}{\code{named.list()|named.atomic()}

A named list or atomic vector}
}
\description{
It returns a logical mask indicating which rows have all columns in the list equal to the value in the list. It only
calculates equality with a reducing function \code{&}.
}
\details{
Any attempt to generalize this function is very likely a mistake as it is thought to be a shorthand
for a common operation.
}
\keyword{internal}
