% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_forest.R
\name{forest_server}
\alias{forest_server}
\title{Forest plot server function}
\usage{
forest_server(
  id,
  bm_dataset,
  group_dataset,
  dataset_name = shiny::reactive(character(0)),
  numeric_numeric_functions = list(),
  numeric_factor_functions = list(),
  subjid_var = "USUBJID",
  cat_var = "PARCAT",
  par_var = "PARAM",
  visit_var = "AVISIT",
  value_vars = "AVAL",
  default_cat = NULL,
  default_par = NULL,
  default_visit = NULL,
  default_value = NULL,
  default_var = NULL,
  default_group = NULL,
  default_categorical_A = NULL,
  default_categorical_B = NULL
)
}
\arguments{
\item{id}{\verb{[character(1)]}

Shiny ID}

\item{bm_dataset}{\verb{[data.frame()]}

An ADBM-like dataset similar in structure to the one in
\href{https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192}{this example},
with one record per subject per parameter per analysis visit.

It should have, at least, the columns specified by the parameters \code{subjid_var}, \code{cat_var},
\code{par_var}, \code{visit_var} and \code{value_vars}.
The semantics of these columns are as described in the CDISC standard for variables
USUBJID, PARCAT, PARAM, AVISIT and AVAL, respectively.}

\item{group_dataset}{\verb{[data.frame()]}

An ADSL-like dataset similar in structure to the one in
\href{https://www.cdisc.org/kb/examples/adam-subject-level-analysis-adsl-dataset-80283806}{this example},
with one record per subject.

It should contain, at least, the column specified by the parameter \code{subjid_var}.}

\item{dataset_name}{\verb{[shiny::reactive(*)]}

A reactive that indicates a possible change in the column structure of any of the two datasets}

\item{numeric_numeric_functions, numeric_factor_functions}{\verb{[function(n)]}

Named lists of functions. Each function needs to take two parameters and produce a list of four numbers
with the following names:
\itemize{
\item result, CI_lower_limit, CI_upper_limit and p_value
}

The module will offer the functions as part of its interface and will run each function if selected.

The values returned by the functions are be captured on the output table and are also displayed
as part of the forest plot.

\code{numeric_numeric_functions} take two numeric parameters (e.g. \code{dv.explorer.parameter::pearson_correlation})
and \code{numeric_factor_functions} should accept a numeric first parameter and a categorical (factor) second parameter
(e.g. \code{dv.explorer.parameter::odds_ratio}).}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to the subject ID}

\item{cat_var, par_var, visit_var}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{default_cat, default_par, default_visit, default_value}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_var, default_group, default_categorical_A, default_categorical_B}{\verb{[character(1)|NULL]}

Default values for the selectors}
}
\description{
Forest plot server function
}
\keyword{developers}
