% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_forest.R
\name{forest_UI}
\alias{forest_UI}
\title{Forest plot UI function}
\usage{
forest_UI(
  id,
  numeric_numeric_function_names = character(0),
  numeric_factor_function_names = character(0),
  default_function = NULL
)
}
\arguments{
\item{id}{\verb{[character(1)]}

Shiny ID}

\item{numeric_numeric_function_names, numeric_factor_function_names}{\verb{[character(1)]}

Vectors of names of functions passed as \code{numeric_numeric_functions} and \code{numeric_factor_functions}
to \code{forest_server}}

\item{default_function}{\verb{[character(1)]}

Default function}
}
\description{
Forest plot UI function
}
\keyword{developers}
