% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_forest.R
\name{fp_stats}
\alias{fp_stats}
\alias{pearson_correlation}
\alias{spearman_correlation}
\alias{odds_ratio}
\title{Forest plot comnputing functions}
\usage{
pearson_correlation(a, b)

spearman_correlation(a, b)

odds_ratio(a, b)
}
\arguments{
\item{a, b}{vectors to be tested}
}
\description{
Helpers used in the forest plot to calculate statistics
}
\section{Functions}{
\itemize{
\item \code{pearson_correlation()}: Pearson correlation function

Forest plot helper

\item \code{spearman_correlation()}: Spearman correlation function

Forest plot helper

\item \code{odds_ratio()}: Odds ratio function

Forest plot helper

}}
