% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_forest.R
\name{fp_subset_data}
\alias{fp_subset_data}
\title{Subset datasets for correlation heatmap}
\usage{
fp_subset_data(
  cat,
  cat_col,
  par,
  par_col,
  val_col,
  vis,
  vis_col,
  group_vect,
  bm_ds,
  group_ds,
  subj_col
)
}
\arguments{
\item{par, cat}{\verb{[character*ish*(n)]}

Values from \code{par_col} and \code{cat_col} to be subset}

\item{val_col}{\verb{[character*ish*(1)]}

Column containing values for the parameters}

\item{vis}{\verb{[character*ish*(1)]}

Values from \code{vis_col} to be subset}

\item{group_vect}{\verb{[named(character(n))]}

Columns to be subset and renamed.}

\item{bm_ds, group_ds}{\code{data.frame}

data frames to be used as inputs in \link{subset_bds_param} and \link{subset_adsl}}

\item{subj_col, par_col, cat_col, vis_col}{\verb{[character(1)]}

Column for subject id, category, parameter and visit. All specified columns must be factors}
}
\description{
TODO: Update this roxygen description to match the function
}
\keyword{internal}
