% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_dens_data}
\alias{get_dens_data}
\title{Calculate probability density per group}
\usage{
get_dens_data(ds)
}
\arguments{
\item{ds}{\verb{[data.frame()]}
A dataframe}
}
\value{
\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_value} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{dens_x} \verb{[numeric()]}: Predictor parameter value.
\item \code{dens_y} \verb{[numeric()]}: Predictor parameter probability density.
}
}
\description{
It calculates the probability density of the parameter value by groups defined by \code{predictor_parameter},
\code{response_parameter} and, if grouped, \code{group}.
These functions is prepared to be applied over a dataset that is the output of \code{roc_subset_data()}.
}
\details{
\link[stats:density]{stats::density} with default values is used to calculate the density
}
\section{Input dataframe}{


\verb{[data.frame()]}

With columns:
\itemize{
\item \code{subject_id} \verb{[factor()]}: Subject ID
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{predictor_value} \verb{[numeric()]}: Predictor parameter Value
\item \code{response_value} \verb{[factor()]}: Response parameter Value.
}
}

\keyword{internal}
