% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_dens_spec}
\alias{get_dens_spec}
\title{Specification for a set of faceted probability density plots}
\usage{
get_dens_spec(ds, param_as_cols, fig_size)
}
\arguments{
\item{ds}{\verb{[data.frames()]}

A dataframe}

\item{param_as_cols}{\verb{[logical(1)]}

Charts are faceted using parameters as columns.}

\item{fig_size}{\verb{[numeric(1)]}

Size in pixels for each of the charts in the facet}
}
\value{
A \link{vegawidget} specification
}
\description{
The chart consists of a faceted set of ROC curves plotting probability density curves for the values of
the different predictor parameters. Each facet will be grouped by the value of the response parameter.
}
\details{
\subsection{Rows and columns:}{
\itemize{
\item By default, predictor parameters are displayed in rows while grouping is displayed in columns.
\item If no grouping is selected, then parameters are displayed in cols.
\item All groups corresponding to the same parameters share the same X and Y axis limits.
}
}
}
\section{Input dataframe}{


\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_value} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{dens_x} \verb{[numeric()]}: Predictor parameter value.
\item \code{dens_y} \verb{[numeric()]}: Predictor parameter probability density.
}
}

\keyword{internal}
