% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_explore_roc_data}
\alias{get_explore_roc_data}
\title{Prepare dataframe for auc exploration}
\usage{
get_explore_roc_data(ds)
}
\arguments{
\item{ds}{\verb{[data.frame()]}}
}
\value{
\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{auc} \verb{[numeric()]}: Area under the curve
\item \code{lower_CI_auc} \verb{[numeric()]}: AUC lower confidence interval
\item \code{upper_CI_auc} \verb{[numeric()]}: AUC upper confidence interval
}
}
\description{
Removes sensitivity, specificity, and threshold and expands auc in three columns.
Removes duplicated rows.
Used on the output of \code{\link[=get_roc_data]{get_roc_data()}}
}
\details{
It respects columns labels
}
\section{Input dataframe}{


\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{specificity} \verb{[numeric()]}: Sensitivity
\item \code{sensitivity} \verb{[numeric()]}: Specificity
\item \code{threshold} \verb{[numeric()]}: Threshold
\item \code{auc} \verb{[numeric(3)]}: A numeric vector of length 3 c(LOWER AUC CI, AUC, UPPER AUC CI)
}
}

\keyword{internal}
