% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_gt_summary_table}
\alias{get_gt_summary_table}
\title{GT Summary table for the area under the curve, optimal cut data and their confidence intervals}
\usage{
get_gt_summary_table(ds, rounder = function(x) round(x, digits = 2), sort_alph)
}
\arguments{
\item{ds}{\verb{[data.frames()]}

A dataframe}

\item{rounder}{\verb{[function()]}

Single parameter function that rounds a numeric vector for rendendering}

\item{sort_alph}{\verb{[logical(1)]}

Sort summary table by parameter name}
}
\value{
A \link{vegawidget} specification
}
\description{
The table summarizes the AUC, the optimal cut data, for each parameter and grouping combination
}
\details{
Numerical values are rounded to two decimals
}
\section{Input}{


\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{auc} \verb{[numeric()]}: Area under the curve
\item \code{lower_CI_auc} \verb{[numeric()]}: AUC lower confidence interval
\item \code{upper_CI_auc} \verb{[numeric()]}: AUC upper confidence interval
\item \code{optimal_cut_title} \verb{[character()]}: Name of the optimal cut
\item \code{optimal_cut_specificity} \verb{[numeric()]}: Sensitivity at the optimal cut point
\item \code{optimal_cut_lower_specificity} \verb{[numeric()]}: Lower Confidence interval of sensitivity
\item \code{optimal_cut_upper_specificity} \verb{[numeric()]}: Upper Confidence interval of sensitivity
\item \code{optimal_cut_sensitivity} \verb{[numeric()]}: Specificity at the optimal cut point
\item \code{optimal_cut_lower_sensitivity} \verb{[numeric()]}: Lower Confidence interval of sensitivity
\item \code{optimal_cut_upper_sensitivity} \verb{[numeric()]}: Upper Confidence interval of sensitivity
}
}

\keyword{internal}
