% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_histo_data}
\alias{get_histo_data}
\title{Calculate binned count per group}
\usage{
get_histo_data(ds)
}
\arguments{
\item{ds}{\verb{[data.frame()]}
A dataframe}
}
\value{
\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_value} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{bin_start} \verb{[numeric()]}: Predictor parameter bin start.
\item \code{bin_end} \verb{[numeric()]}: Predictor parameter bin end.
\item \code{bin_count} \verb{[numeric()]}: Predictor parameter bin count.
}
}
\description{
It calculates the binned count of the parameter value by groups defined by \code{predictor_parameter}, \code{response_parameter} and
if grouped, \code{group}. These functions is prepared to be applied over a dataset that is the output of
\code{subset_data()}.
}
\details{
\link[graphics:hist]{graphics::hist} with default values is used to calculate the binning
}
\section{Input dataframe}{


\verb{[data.frame()]}

With columns:
\itemize{
\item \code{subject_id} \verb{[factor()]}: Subject ID
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{predictor_value} \verb{[numeric()]}: Predictor parameter Value
\item \code{response_value} \verb{[factor()]}: Response parameter Value.
}
}

\keyword{internal}
