% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_metric_data}
\alias{get_metric_data}
\title{Compute classification metrics to groups}
\usage{
get_metric_data(ds, compute_metric_fn)
}
\arguments{
\item{ds}{\verb{[data.frame()]}
A dataframe}

\item{compute_metric_fn}{\verb{[function(1)]}
A function that computes the metrics}
}
\value{
\verb{[list()]}
A list with entries:
\subsection{\code{ds}}{

\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{type} \verb{[character()]}: Metric name
\item \code{y} \verb{[numeric()]}: Metric value
\item \code{score}, \code{norm_score}, \code{norm_rank} \verb{[numeric()]}: Raw, normalized, and normalized rank predictor value per group.
}
}

\subsection{\code{limits}}{

\verb{[list()]}

With one entry per metric type. Each entry is a \code{numeric(2)} that contains the plotting limits for the metric.

Additionally:
\itemize{
\item \code{score}, \code{norm_score}, \code{norm_rank} have a \code{label} attribute \code{"Score"}, \code{"Normalized Score"}
and \code{"Normalized Rank"} respectively.
}
}
}
\description{
It computes classification metrics over a dataset. The application is applied by groups defined by CNT_ROC$PPAR,
CNT_ROC$RPAR and, if grouped, CNT_ROC$GRP. These functions is prepared to be applied over a dataset that is the
output of \code{roc_subset_data()}. The function itself does not compute the metrics, it only applies \code{compute_metric_fn}
over the different groups.
}
\section{Input dataframe}{


\verb{[data.frame()]}

With columns:
\itemize{
\item \code{subject_id} \verb{[factor()]}: Subject ID
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{predictor_value} \verb{[numeric()]}: Predictor parameter Value
\item \code{response_value} \verb{[factor()]}: Response parameter Value.
}
}

\section{compute_metric_fn definition}{


For an example of a computing function please review \code{\link[=compute_metric_data]{compute_metric_data()}}.
}

\keyword{internal}
