% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_metric_spec}
\alias{get_metric_spec}
\title{Specification for a set of line plots for classification metrics}
\usage{
get_metric_spec(ds, param_as_cols, fig_size, limits, x_col)
}
\arguments{
\item{ds}{\verb{[data.frames()]}

A dataframe}

\item{param_as_cols}{\verb{[logical(1)]}

Charts are faceted using parameters as columns.}

\item{fig_size}{\verb{[numeric(1)]}

Size in pixels for each of the charts in the facet}

\item{limits}{\verb{[list()]}

A vector as specified in the input section}

\item{x_col}{\verb{[character(1)]}

Column for the x axis of the graph}
}
\value{
A \link{vegawidget} specification
}
\description{
The chart consists of a faceted set of line and point charts.
}
\details{
\subsection{Rows and columns:}{
\itemize{
\item By default, predictor parameters are displayed in rows while metrics are displayed in columns.
}
}
}
\section{Input}{

\subsection{\code{ds}}{

\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{type} \verb{[character()]}: Metric name
\item \code{y} \verb{[numeric()]}: Metric value
\item \code{score}, \code{norm_score}, \code{norm_rank} \verb{[numeric()]}: Raw, normalized, and normalized rank predictor value per group.
}
}

\subsection{\code{limits}}{

\verb{[list()]}

With one entry per metric type. Each entry is a \code{numeric(2)} that contains the plotting limits for the metric.
}
}

\keyword{internal}
