% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_quantile_data}
\alias{get_quantile_data}
\title{Calculate quantiles per group}
\usage{
get_quantile_data(ds)
}
\arguments{
\item{ds}{\verb{[data.frame()]}
A dataframe}
}
\value{
\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_value} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{mean} \verb{[numeric()]}: Mean of the predictor parameter value
\item \code{q05}, \code{q25}, \code{q50}, \code{q75}, \code{q95} \verb{[numeric()]}: Quantiles of the predictor parameter value
}
}
\description{
It calculates the quantiles \code{c(.05, .25, .50, .75, .95)} and the mean of the parameter value by groups defined by
CNT_ROC$PPAR, CNT_ROC$RPAR and, if grouped, CNT_ROC$GRP. These functions is prepared to be applied over a dataset
that is the output of \code{roc_subset_data()}.
}
\section{Input dataframe}{


\verb{[data.frame()]}

With columns:
\itemize{
\item \code{subject_id} \verb{[factor()]}: Subject ID
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{predictor_value} \verb{[numeric()]}: Predictor parameter Value
\item \code{response_value} \verb{[factor()]}: Response parameter Value.
}
}

\keyword{internal}
