% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_roc_spec}
\alias{get_roc_spec}
\alias{get_roc_sorted_spec}
\title{Specification for a set of faceted ROC curves}
\usage{
get_roc_spec(ds_list, param_as_cols, fig_size)

get_roc_sorted_spec(ds_list, param_as_cols, fig_size)
}
\arguments{
\item{ds_list}{\verb{[list(data.frames())]}

A list of dataframes}

\item{param_as_cols}{\verb{[logical(1)]}

Charts are faceted using parameters as columns. This parameter is ignored in the sorted version.}

\item{fig_size}{\verb{[numeric(1)]}

Size in pixels for each of the charts in the facet}
}
\value{
A \link{vegawidget} specification
}
\description{
The chart consists of a faceted set of ROC curves plotting the:
}
\details{
\itemize{
\item ROC curves
\item The confidence intervals
\item The set of optimal cuts
}
\subsection{Rows and columns:}{
\itemize{
\item By default, predictor parameters are displayed in rows while grouping is displayed in columns.
\item If no grouping is selected, then parameters are displayed in cols.
\item If \verb{ds_list[[}r CNT_ROC$ROC_CI]]`` is \code{NULL} then confidence intervals are not presented.
}
}

\subsection{Rows and columns:}{
\itemize{
\item Charts are ordered from left to right and top to bottom from the highest to lowest area under the curve
}
}
}
\section{Functions}{
\itemize{
\item \code{get_roc_sorted_spec()}: Specification for a set of sorted ROC curves

}}
\section{Input dataframe list}{

\subsection{\code{ds_list}}{

\verb{[list(data.frame())]}
A list with entries:
}

\subsection{roc_curve}{

\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{specificity} \verb{[numeric()]}: Sensitivity
\item \code{sensitivity} \verb{[numeric()]}: Specificity
\item \code{threshold} \verb{[numeric()]}: Threshold
\item \code{auc} \verb{[numeric(3)]}: A numeric vector of length 3 c(LOWER AUC CI, AUC, UPPER AUC CI)
}
}

\subsection{roc_ci}{

\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{ci_specificity} \verb{[numeric()]}: Specificity value
\item \code{ci_lower_specificity} \verb{[numeric()]}: Specificity lower confidence interval
\item \code{ci_upper_specificity} \verb{[numeric()]}: Specificity upper confidence interval
\item \code{ci_sensitivity} \verb{[numeric()]}: Sensitivity value
\item \code{ci_lower_sensitivity} \verb{[numeric()]}: Sensitivity lower confidence interval
\item \code{ci_upper_sensitivity} \verb{[numeric()]}: Sensitivity upper confidence interval
}

CIs are only calculated when \code{do_bootstrap} is \code{TRUE}
}

\subsection{roc_optimal_cut}{

\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{optimal_cut_title} \verb{[character()]}: Name of the optimal cut
\item \code{optimal_cut_specificity} \verb{[numeric()]}: Sensitivity at the optimal cut point
\item \code{optimal_cut_lower_specificity} \verb{[numeric()]}: Lower Confidence interval of sensitivity
\item \code{optimal_cut_upper_specificity} \verb{[numeric()]}: Upper Confidence interval of sensitivity
\item \code{optimal_cut_sensitivity} \verb{[numeric()]}: Specificity at the optimal cut point
\item \code{optimal_cut_lower_sensitivity} \verb{[numeric()]}: Lower Confidence interval of sensitivity
\item \code{optimal_cut_upper_sensitivity} \verb{[numeric()]}: Upper Confidence interval of sensitivity
\item \code{optimal_cut_threshold} \verb{[numeric()]}: Threshold of the optimal cut
}

CIs are only calculated when \code{do_bootstrap} is \code{TRUE}
}
}

\keyword{internal}
