% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{get_summary_data}
\alias{get_summary_data}
\title{Prepare dataframe for summary table exploration}
\usage{
get_summary_data(ds_list)
}
\arguments{
\item{ds_list}{\verb{[list(data.frame)]}}
}
\value{
\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{auc} \verb{[numeric()]}: Area under the curve
\item \code{lower_CI_auc} \verb{[numeric()]}: AUC lower confidence interval
\item \code{upper_CI_auc} \verb{[numeric()]}: AUC upper confidence interval
\item \code{optimal_cut_title} \verb{[character()]}: Name of the optimal cut
\item \code{optimal_cut_specificity} \verb{[numeric()]}: Sensitivity at the optimal cut point
\item \code{optimal_cut_lower_specificity} \verb{[numeric()]}: Lower Confidence interval of sensitivity
\item \code{optimal_cut_upper_specificity} \verb{[numeric()]}: Upper Confidence interval of sensitivity
\item \code{optimal_cut_sensitivity} \verb{[numeric()]}: Specificity at the optimal cut point
\item \code{optimal_cut_lower_sensitivity} \verb{[numeric()]}: Lower Confidence interval of sensitivity
\item \code{optimal_cut_upper_sensitivity} \verb{[numeric()]}: Upper Confidence interval of sensitivity
}
}
\description{
Expands auc from roc_curve in three columns and joins the dataset with roc_optimal_cut
Used on the output of \code{\link[=get_roc_data]{get_roc_data()}}
}
\section{Input dataframe list}{


\verb{[list(data.frame())]}
A list with entries:
\subsection{roc_curve}{

\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{specificity} \verb{[numeric()]}: Sensitivity
\item \code{sensitivity} \verb{[numeric()]}: Specificity
\item \code{threshold} \verb{[numeric()]}: Threshold
\item \code{auc} \verb{[numeric(3)]}: A numeric vector of length 3 c(LOWER AUC CI, AUC, UPPER AUC CI)
}
}

\subsection{roc_ci}{

\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{ci_specificity} \verb{[numeric()]}: Specificity value
\item \code{ci_lower_specificity} \verb{[numeric()]}: Specificity lower confidence interval
\item \code{ci_upper_specificity} \verb{[numeric()]}: Specificity upper confidence interval
\item \code{ci_sensitivity} \verb{[numeric()]}: Sensitivity value
\item \code{ci_lower_sensitivity} \verb{[numeric()]}: Sensitivity lower confidence interval
\item \code{ci_upper_sensitivity} \verb{[numeric()]}: Sensitivity upper confidence interval
}

CIs are only calculated when \code{do_bootstrap} is \code{TRUE}
}

\subsection{roc_optimal_cut}{

\verb{[data.frame()]}

With columns:
\itemize{
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter name.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{optimal_cut_title} \verb{[character()]}: Name of the optimal cut
\item \code{optimal_cut_specificity} \verb{[numeric()]}: Sensitivity at the optimal cut point
\item \code{optimal_cut_lower_specificity} \verb{[numeric()]}: Lower Confidence interval of sensitivity
\item \code{optimal_cut_upper_specificity} \verb{[numeric()]}: Upper Confidence interval of sensitivity
\item \code{optimal_cut_sensitivity} \verb{[numeric()]}: Specificity at the optimal cut point
\item \code{optimal_cut_lower_sensitivity} \verb{[numeric()]}: Lower Confidence interval of sensitivity
\item \code{optimal_cut_upper_sensitivity} \verb{[numeric()]}: Upper Confidence interval of sensitivity
\item \code{optimal_cut_threshold} \verb{[numeric()]}: Threshold of the optimal cut
}

CIs are only calculated when \code{do_bootstrap} is \code{TRUE}
}
}

\keyword{internal}
