% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineplot.R
\name{lineplot_server}
\alias{lineplot_server}
\title{Lineplot server function}
\usage{
lineplot_server(
  id,
  bm_dataset,
  group_dataset,
  summary_fns = list(Mean = lp_mean_summary_fns, Median = lp_median_summary_fns),
  subjid_var = "USUBJID",
  cat_var = "PARCAT",
  par_var = "PARAM",
  visit_vars = c("AVISIT"),
  cdisc_visit_vars = character(0),
  value_vars = "AVAL",
  additional_listing_vars = character(0),
  ref_line_vars = character(0),
  on_sbj_click = NULL,
  default_centrality_fn = NULL,
  default_dispersion_fn = NULL,
  default_cat = NULL,
  default_par = NULL,
  default_val = NULL,
  default_visit_var = NULL,
  default_visit_val = NULL,
  default_main_group = NULL,
  default_sub_group = NULL,
  default_transparency = 1,
  default_y_axis_projection = "Linear"
)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{bm_dataset}{\verb{[data.frame()]}

An ADBM-like dataset similar in structure to the one in
\href{https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192}{this example},
with one record per subject per parameter per analysis visit.

It should have, at least, the columns specified by the parameters \code{subjid_var}, \code{cat_var},
\code{par_var}, \code{visit_vars} and \code{value_vars}.
The semantics of these columns are as described in the CDISC standard for variables
USUBJID, PARCAT, PARAM, AVISIT and AVAL, respectively.

Optional columns specified by \code{ref_line_vars} should contain the same numeric value for all
records of the same parameter for any given subject.}

\item{group_dataset}{\verb{[data.frame()]}

An ADSL-like dataset similar in structure to the one in
\href{https://www.cdisc.org/kb/examples/adam-subject-level-analysis-adsl-dataset-80283806}{this example},
with one record per subject.

It should contain, at least, the column specified by the parameter \code{subjid_var}.}

\item{summary_fns}{\verb{[list()]}

Each element of this named list contains a summary function (e.g. a mean) and a collection of dispersion functions
(e.g. standard deviation) defining ranges around the values returned by the summary function.

The structure of each element is then a named list with the following elements:
\itemize{
\item \code{fn}: Function that takes a numeric vector as its sole parameter and produces a scalar number.
\item \code{dispersion}: Named list of pairs functions that return the \emph{top} and \emph{bottom} dispersion ranges. They also take
a numeric vector as input and return a single numeric scalar
\item \code{y_prefix}: Prefix that will be prepended the Y axis label of the generated plot
}

For an example, see \code{dv.explorer.parameter::lp_mean_summary_fns}.}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to the subject ID}

\item{cat_var, par_var, visit_vars}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{cdisc_visit_vars}{\verb{[character(1)]}

Column from \code{bm_dataset} that correspond to the parameter visit and is interpreted as a CDISC
Visit Days (skipping day 0; jumping from value -1 to value 1 in the X axis)}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{additional_listing_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that will be appended to the single-subject listing}

\item{ref_line_vars}{\verb{[character(n)]}

Columns for \code{bm_dataset} specifying reference values for parameters.
See \href{../articles/lineplot_reference_values.html}{this article} for more details}

\item{on_sbj_click}{\verb{[function()]}

Function to invoke when a subject is clicked in the single-subject listing}

\item{default_cat, default_par, default_visit_var, default_main_group}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_visit_val}{\verb{list([character(n)|numeric(n)])}

Named list of default values associated to specific \code{visit_var}s, e.g.
\code{default_visit_val = list(VISIT = c('VISIT1', 'VISIT2'), AVISITN = c(1, 2))}}

\item{default_sub_group, default_val, default_centrality_fn, default_dispersion_fn}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_transparency}{\verb{[numeric(1)]}

Default values for the selectors}

\item{default_y_axis_projection}{\verb{["Linear"|"Logarithmic"]}

Default projection for the Y axis}
}
\description{
Lineplot server function
}
\keyword{developers}
