% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineplot.R
\name{lineplot_UI}
\alias{lineplot_UI}
\title{Lineplot UI function}
\usage{
lineplot_UI(id)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}
}
\description{
Lineplot UI function
}
\keyword{developers}
