% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_boxplot.R
\name{mock_app_boxplot_mm_depr}
\alias{mock_app_boxplot_mm_depr}
\title{Mock mm boxplot app with deprecated visit_var and default_visit arguments}
\usage{
mock_app_boxplot_mm_depr(update_query_string = TRUE, anlfl_flags = FALSE)
}
\arguments{
\item{update_query_string}{automatically update query string with app state}

\item{anlfl_flags}{indicates that the input data contain analysis flag variables or not}
}
\description{
Mock mm boxplot app with deprecated visit_var and default_visit arguments
}
\keyword{mock}
