% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_corr_hm.R
\name{mock_app_corr_hm}
\alias{mock_app_corr_hm}
\title{Mock corr hm app}
\usage{
mock_app_corr_hm(
  dry_run = FALSE,
  update_query_string = TRUE,
  srv_defaults = list(),
  ui_defaults = list()
)
}
\arguments{
\item{dry_run}{Return parameters used in the call}

\item{update_query_string}{automatically update query string with app state}

\item{ui_defaults, srv_defaults}{a list of values passed to the ui/server function}
}
\description{
Mock corr hm app
}
\keyword{mock}
