% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_corr_hm.R
\name{mock_app_correlation_hm_mm}
\alias{mock_app_correlation_hm_mm}
\title{Mock corr hm app}
\usage{
mock_app_correlation_hm_mm(anlfl_flags = FALSE)
}
\arguments{
\item{anlfl_flags}{indicates that the input data contain analysis flag variables or not}
}
\description{
Mock corr hm app
}
\keyword{mock}
