% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_lineplot.R
\name{mock_app_lineplot_mm_safetyData}
\alias{mock_app_lineplot_mm_safetyData}
\title{Mock module manager lineplot app displaying safetyData dataset}
\usage{
mock_app_lineplot_mm_safetyData()
}
\description{
Mock module manager lineplot app displaying safetyData dataset
}
\keyword{mock}
