% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_scatter.R
\name{mock_app_scatterplot}
\alias{mock_app_scatterplot}
\title{Mock scatterplot app}
\usage{
mock_app_scatterplot(
  dry_run = FALSE,
  update_query_string = TRUE,
  srv_defaults = list(),
  ui_defaults = list(),
  anlfl_flags = FALSE
)
}
\arguments{
\item{dry_run}{Return parameters used in the call}

\item{update_query_string}{automatically update query string with app state}

\item{ui_defaults, srv_defaults}{a list of values passed to the ui/server function}

\item{anlfl_flags}{indicates that the input data contain analysis flag variables or not}
}
\description{
Mock scatterplot app
}
\keyword{mock}
