% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{mock_roc}
\alias{mock_roc}
\alias{mock_roc_mm_app}
\alias{mock_roc_app}
\title{Mock functions}
\usage{
mock_roc_mm_app(
  adbm = test_roc_data()[["adbm"]],
  adbin = test_roc_data()[["adbin"]],
  group = test_roc_data()[["adsl"]]
)

mock_roc_app()
}
\arguments{
\item{adbm, adbin, group}{Datasets for the mock app}
}
\description{
Mock functions
}
\section{Functions}{
\itemize{
\item \code{mock_roc_mm_app()}: Mock app running the module inside dv.manager

\item \code{mock_roc_app()}: Mock app running the module standalone

}}
\keyword{mock}
