% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_wfphm.R
\name{mock_wfphm}
\alias{mock_wfphm}
\alias{mock_wfphm_mm_app}
\alias{mock_app_wfphm2}
\alias{mock_app_hmcat}
\alias{mock_app_hmcont}
\alias{mock_app_hmpar}
\alias{mock_app_wf}
\alias{mock_app_wfphm}
\title{Mock functions}
\usage{
mock_wfphm_mm_app()

mock_app_wfphm2(
  dry_run = FALSE,
  update_query_string = TRUE,
  srv_defaults = list(),
  ui_defaults = list()
)

mock_app_hmcat(
  dry_run = FALSE,
  update_query_string = TRUE,
  srv_args = list(),
  ui_args = list()
)

mock_app_hmcont(
  dry_run = FALSE,
  update_query_string = TRUE,
  srv_args = list(),
  ui_args = list()
)

mock_app_hmpar(
  dry_run = FALSE,
  update_query_string = TRUE,
  srv_args = list(),
  ui_args = list()
)

mock_app_wf(
  dry_run = FALSE,
  update_query_string = TRUE,
  srv_args = list(),
  ui_args = list()
)

mock_app_wfphm(
  dry_run = FALSE,
  update_query_string = TRUE,
  srv_args = list(),
  ui_args = list()
)
}
\arguments{
\item{dry_run}{Return parameters used in the call}

\item{update_query_string}{automatically update query string with app state}

\item{ui_defaults, srv_defaults}{a list of values passed to the ui/server function}

\item{ui_args, srv_args}{a list of arguments passed to the ui/server function.}
}
\description{
Mock functions
}
\section{Functions}{
\itemize{
\item \code{mock_wfphm_mm_app()}: Mock app running the module inside dv.manager

\item \code{mock_app_wfphm2()}: Mock app running the waterfall plus heatmap module

\item \code{mock_app_hmcat()}: Mock app running categorical heatmap module

\item \code{mock_app_hmcont()}: Mock app running continuous heatmap module

\item \code{mock_app_hmpar()}: Mock app running parameter heatmap module

\item \code{mock_app_wf()}: Mock app running waterfall module

\item \code{mock_app_wfphm()}: Mock app running the waterfall plus heatmap module

}}
\keyword{mock}
