% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{mod_boxplot}
\alias{mod_boxplot}
\title{Boxplot module}
\usage{
mod_boxplot(
  module_id,
  bm_dataset_name,
  group_dataset_name,
  receiver_id = NULL,
  cat_var = "PARCAT",
  par_var = "PARAM",
  value_vars = "AVAL",
  x_axis_vars = NULL,
  visit_var = NULL,
  anlfl_vars = NULL,
  subjid_var = "SUBJID",
  quantile_type = 7L,
  default_cat = NULL,
  default_par = NULL,
  default_x_axis_var = NULL,
  default_x_axis_vals = NULL,
  default_visit = NULL,
  default_value = NULL,
  default_main_group = NULL,
  default_sub_group = NULL,
  default_page_group = NULL,
  server_wrapper_func = function(x) list(subj_id = x)
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

Module Shiny id}

\item{bm_dataset_name, group_dataset_name}{\verb{[character(1)]}

Name of the dataset}

\item{receiver_id}{\verb{[character(1)]}

Shiny ID of the module receiving the selected subject ID in the data listing. This ID must
be present in the app or be NULL.}

\item{cat_var, par_var}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category and parameter}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{x_axis_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to x-axis variables}

\item{visit_var}{Deprecated. Use \code{x_axis_vars}.}

\item{anlfl_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to analysis flags}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to subject ID}

\item{quantile_type}{\verb{[integer(1)]}

Quantile algorithm type passed to \code{\link[stats]{quantile}}
(an integer between 1 and 9, default 7).}

\item{default_cat, default_par, default_x_axis_var, default_value, default_main_group, default_sub_group, default_page_group}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_x_axis_vals}{\verb{[character(n)|NULL]}

Default values for the x-axis variable}

\item{default_visit}{Deprecated. Use \code{default_x_axis_vals}.}

\item{server_wrapper_func}{\verb{[function()]}

A function that will be applied to the server returned value. Its default value will work for the current cases.}
}
\description{
\code{mod_boxplot} is a Shiny module prepared to display data with boxplot charts with different levels of grouping.
It also includes a set of listings with information about the population, distribution and statistical comparisons.
}
\keyword{main}
