% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{mod_boxplot_papo}
\alias{mod_boxplot_papo}
\title{Boxplot wrapper when its output is fed into the papo module}
\usage{
mod_boxplot_papo(...)
}
\description{
Boxplot wrapper when its output is fed into the papo module
}
\keyword{main}
