% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter.R
\name{mod_scatterplot}
\alias{mod_scatterplot}
\title{Scatterplot module}
\usage{
mod_scatterplot(
  module_id,
  bm_dataset_name,
  group_dataset_name,
  cat_var = "PARCAT",
  par_var = "PARAM",
  value_vars = "AVAL",
  visit_var = "AVISIT",
  ref_line_vars = character(0),
  anlfl_vars = NULL,
  subjid_var = "USUBJID",
  default_x_cat = NULL,
  default_x_par = NULL,
  default_x_value = NULL,
  default_x_visit = NULL,
  default_y_cat = NULL,
  default_y_par = NULL,
  default_y_value = NULL,
  default_y_visit = NULL,
  default_main_group = NULL,
  default_sub_group = NULL,
  default_group = NULL,
  default_color = NULL,
  default_include_regression_info = TRUE,
  compute_lm_cor_fn = sp_compute_lm_cor_default
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

Module Shiny id}

\item{bm_dataset_name, group_dataset_name}{\verb{[character(1)]}

Name of the dataset}

\item{cat_var, par_var, visit_var}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{ref_line_vars}{\verb{[character(n)]}

Columns for \code{bm_dataset} specifying reference values for parameters.
See \href{../articles/scatterplot_reference_values.html}{this article} for more details}

\item{anlfl_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to analysis flags}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to subject ID}

\item{default_x_cat, default_x_par, default_x_visit, default_x_value, default_y_cat, default_y_par}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_y_visit, default_y_value, default_main_group, default_sub_group, default_group, default_color}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{compute_lm_cor_fn}{\verb{[function()]}

Function used to compute the linear regression model and correlation statistics,
please view the corresponding vignette for more details.}
}
\description{
\code{mod_scatterplot} is a Shiny module prepared to display a scatterplot of two biomarkers with different levels of
grouping.
It also includes a set of listings with information about the population and the regression and correlation
estimates.
}
\keyword{main}
