% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imod_outliers.R
\name{outlier_container}
\alias{outlier_container}
\alias{outlier_container_UI}
\alias{outlier_container_server}
\alias{mock_outlier}
\title{A container for outlier selectors}
\usage{
outlier_container_UI(id)

outlier_container_server(id, choices)

mock_outlier()
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{choices}{a list with the parameters for which we want to select outliers}
}
\value{
A reactive containing a list with as many entries as choices are selected in the selector. Each of the
entries contain a list with two entries\code{ll} and \code{ul} as described in \link{outlier_selector}.
}
\description{
A container for outlier selectors
}
\details{
When a choice is deselected and selected again, it will retain its previous state.
}
\section{Functions}{
\itemize{
\item \code{outlier_container_UI()}: UI

\item \code{outlier_container_server()}: server

\item \code{mock_outlier()}: mock

}}
\seealso{
outlier_selector
}
\keyword{internal}
