% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imod_outliers.R
\name{outlier_selector}
\alias{outlier_selector}
\alias{outlier_ui}
\alias{outlier_server}
\title{Single outlier selector}
\usage{
outlier_ui(id, label, value_ll = "", value_ul = "")

outlier_server(id)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{label}{label for the selectors}

\item{value_ll}{an initial value for the lower limit}

\item{value_ul}{an initial value for the upper limit}
}
\value{
A reactive containing a list with two entries\code{ll} and \code{ul}. It returns either NA if the box is empty or the
result of applying as.numeric to the contents of the input box. It does not return NULL values.
}
\description{
A selector composed by two rows containing:
\itemize{
\item a \code{label}
\item two inline \link[shiny:textInput]{shiny::textInput} boxes for the upper and lower limit of the outliers
}
}
\details{
The UI root element is a divisor which id attribute is equal to the \code{id} parameter. So it can be used
with \link[shiny:insertUI]{shiny::insertUI} and \link[shiny:insertUI]{shiny::removeUI} functions.
}
\section{Functions}{
\itemize{
\item \code{outlier_ui()}: UI

\item \code{outlier_server()}: server

}}
\seealso{
outlier_container
}
\keyword{internal}
