% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{pal_div_palette}
\alias{pal_div_palette}
\title{Returns a divergent palette to use with heatmap}
\usage{
pal_div_palette(max, neut, min, colors)
}
\arguments{
\item{max, neut, min}{the values to be associated with the extreme and central colors}

\item{colors}{the palette colors

max, neut, min the values to be associated with the extreme and central colors. If the palette contains more colors,
the values are interpolated.}
}
\value{
Vector The names of the vector are hexadecimal colors
encoded as #rrggbb or #rrggbbaa (red, green, blue, alpha) and the values are the "z" values that
should be mapped to that color.
}
\description{
Returns a divergent palette to use with heatmap
}
\keyword{internal}
