% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{pal_get_cat_palette}
\alias{pal_get_cat_palette}
\title{Returns a categorical palette to use with heatmap}
\usage{
pal_get_cat_palette(d, colors)
}
\arguments{
\item{d}{a vector containing the values to be colored}

\item{colors}{the colors to be used. They will be recycled as needed.}
}
\value{
Vector The names of the vector are hexadecimal colors
encoded as #rrggbb or #rrggbbaa (red, green, blue, alpha) and the values are the "z" values that
should be mapped to that color.
}
\description{
Returns a categorical palette to use with heatmap
}
\keyword{internal}
