% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{pal_get_cont_palette}
\alias{pal_get_cont_palette}
\title{Returns a continous palette to use with heatmap}
\usage{
pal_get_cont_palette(
  d,
  seq_pos_colors = RColorBrewer::brewer.pal(9, name = "Reds"),
  seq_neg_colors = rev(RColorBrewer::brewer.pal(9, name = "Blues")),
  div_colors = rev(RColorBrewer::brewer.pal(11, name = "RdBu")),
  zero_color = "rgb(242, 239, 238)"
)
}
\arguments{
\item{d}{a vector containing the values to be colored}

\item{seq_pos_colors, seq_neg_colors, div_colors, zero_color}{a palette for sequential positive, sequential
negative, divergent or all zeroes datasets. The zero_color expects a single color, all other expect an odd number of
colors. The divergent palette central color is considered the neutral and is always associated to 0.}
}
\value{
Vector The names of the vector are hexadecimal colors
encoded as #rrggbb or #rrggbbaa (red, green, blue, alpha) and the values are the "z" values that
should be mapped to that color.
}
\description{
Returns a continous palette to use with heatmap
}
\keyword{internal}
