% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{pal_get_scale_type_lim}
\alias{pal_get_scale_type_lim}
\title{Return the type of scale and its limits}
\usage{
pal_get_scale_type_lim(d)
}
\arguments{
\item{d}{the data from which the scale will be inferred}
}
\value{
A list with the entries:
\subsection{type:}{
\itemize{
\item divergent: It has both positive, negative and 0 values
\item seq_positive: It has positive or 0 values
\item seq_negative: It has negative or 0 values
\item seq_positive: It has positive or 0 values
\item all_zero: It has only 0 values
}
}

\subsection{lim:}{
\itemize{
\item If \strong{divergent}: c(-max_abs, 0, max_abs) when max_abs is the maximum absolute value of all data
\item If \strong{seq_positive} or \strong{seq_negative}: c(min(d), max(d)) when max_abs is the maximum absolute value of all data
\item if \strong{all_zerp}: c(0,0)
}
}
}
\description{
Inferes the type of scale and its limits based on the the data
}
\keyword{internal}
