% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{pal_seq_palette}
\alias{pal_seq_palette}
\title{A sequential palette to use with heatmap}
\usage{
pal_seq_palette(max, min, colors)
}
\arguments{
\item{max, min}{the values to be associated with the extreme and colors}

\item{colors}{the palette colors

max, min the values are associated with the extreme colors. If the palette contains more colors,
the values are interpolated.}
}
\value{
Vector The names of the vector are hexadecimal colors
encoded as #rrggbb or #rrggbbaa (red, green, blue, alpha) and the values are the "z" values that
should be mapped to that color.
}
\description{
A sequential palette to use with heatmap
}
\keyword{internal}
