% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-palettes.R
\name{pal_zero_palette}
\alias{pal_zero_palette}
\title{Returns a palette with a single color}
\usage{
pal_zero_palette(color)
}
\arguments{
\item{color}{the color

The palette always return two repeated entries}
}
\value{
Vector The names of the vector are hexadecimal colors
encoded as #rrggbb or #rrggbbaa (red, green, blue, alpha) and the values are the "z" values that
should be mapped to that color.
}
\description{
Returns a palette with a single color
}
\keyword{internal}
