% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{parse_ci}
\alias{parse_ci}
\title{Split string with ; delimiters and transform to numeric}
\usage{
parse_ci(str)
}
\arguments{
\item{str}{\verb{[character(1)]}

String to be split}
}
\description{
Split string with ; delimiters and transform to numeric
}
\keyword{internal}
