% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{prefix_repeat_parameters}
\alias{prefix_repeat_parameters}
\title{Prefix Repeat Parameters}
\usage{
prefix_repeat_parameters(dataset, cat_var, par_var)
}
\arguments{
\item{dataset}{Input data frame.}

\item{cat_var}{Dataset column name that contains the category values.}

\item{par_var}{Dataset column name that contains the parameter values.}
}
\value{
Modified dataset where the repeated parameter values have been prefixed with their corresponding category values.
}
\description{
This function modifies a given dataset by prefixing repeating parameter values with their corresponding category values.
}
\keyword{misc}
