% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_logic.R
\name{rename_with_list}
\alias{rename_with_list}
\title{Rename a data frame with a list/character vector of new names}
\usage{
rename_with_list(ds, rename_vec)
}
\arguments{
\item{ds}{\code{data.frame()}

A dataframe}

\item{rename_vec}{\code{named.list()|named.character()}

A named list or character vector with the new names}
}
\value{
\code{data.frame()}

Returns renamed \code{ds}
}
\description{
The names are the old names and the values are the new names. \code{rename_vec} names not present in \code{ds} are ignored.
}
\keyword{internal}
