% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{resolve_or_return}
\alias{resolve_or_return}
\title{Return the value of a reactive or regular value}
\usage{
resolve_or_return(r)
}
\arguments{
\item{r}{a variable which can be regular or reactive}
}
\value{
the value of the variable or the reactive
}
\description{
Return the value of a reactive or regular value
}
\details{
It does not work with metaReactives.
}
\keyword{internal}
