% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{roc_subset_data}
\alias{roc_subset_data}
\title{Subset input datasets}
\usage{
roc_subset_data(
  pred_cat,
  pred_par,
  pred_val_col,
  pred_visit,
  resp_cat,
  resp_par,
  resp_val_col,
  resp_visit,
  group_col,
  pred_ds,
  resp_ds,
  group_ds,
  subj_col,
  pred_cat_col,
  pred_par_col,
  pred_vis_col,
  resp_cat_col,
  resp_par_col,
  resp_vis_col
)
}
\arguments{
\item{pred_par, pred_cat}{\verb{[character*ish*(n)]}

Values from \code{pred_par_col} and \code{pred_cat_col} to be subset}

\item{pred_val_col}{\verb{[character*ish*(1)]}

Column containing values for the predictor parameters}

\item{pred_visit}{\verb{[character*ish*(1)]}

Values from \code{pred_vis_col} to be subset}

\item{resp_par, resp_cat}{\verb{[character*ish*(1)]}

Values from \code{resp_par_col} and \code{resp_cat_col} to be subset}

\item{resp_val_col}{\verb{[character*ish*(1)]}

Column containing values for the response parameter}

\item{resp_visit}{\verb{[character*ish*(1)]}

Values from \code{resp_vis_col} to be subset}

\item{group_col}{\verb{[character*ish*(1)]}

Column to group the data by. \code{"None"} for no grouping.}

\item{pred_ds, resp_ds, group_ds}{\verb{[data.frame()]}

Data frames for predictors/responses and groupings, see section \emph{Input dataframes}}

\item{subj_col, pred_par_col, pred_cat_col, pred_vis_col, resp_cat_col, resp_par_col, resp_vis_col}{\verb{[character(1)]}

Column for predictor/response category/parameter/visit. \code{subj_col} must be a factor.}
}
\value{
\verb{[data.frame()]}

With columns:
\itemize{
\item \code{subject_id} \verb{[factor()]}: Subject ID
\item \code{predictor_parameter} \verb{[factor()]}: Predictor parameter name.
\item \code{response_parameter} \verb{[factor()]}: Response parameter value.
\item \code{group} \verb{[factor()]}: An optional column for the grouping value (if group is specified).
\item \code{predictor_value} \verb{[numeric()]}: Predictor parameter Value
\item \code{response_value} \verb{[factor()]}: Response parameter Value.
}

Additionally:
\itemize{
\item \code{predictor_parameter} has a label attribute: \emph{Parameter}
\item \code{response_parameter} has a label attribute: \emph{Response}
\item \code{group} has the same label attribute as in the original dataset, if available, otherwise \code{group_col}.
}
}
\description{
This functions prepares the basic input for the rest of dv.explorer.parameter functions.

It subsets and joins the datasets based on the predictor/response category/parameter/visit and group selections.
}
\details{
\itemize{
\item All columns but \code{predictor_value} are coherced into factors if they are not factors already.
\item If at least one parameter name appears under several selected categories,
parameter names and categories will be pasted together.
\item \code{predictor_parameter} is releveled so all extra levels not present in the selection are dropped.
Levels are ordered according to \code{pred_par} as long as a parameter name appears does not appear
under more than one selected categories.
}
}
\section{Input dataframes}{

\subsection{pred_dataset}{

It expects a dataset similar to
https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192 ,
1 record per subject per parameter per analysis visit

It expects, at least, the columns passed in the parameters,
\code{subj_col}, \code{pred_cat_col}, \code{pred_par_col}, \code{pred_vis_col} and \code{pred_val}.
The values of these variables are as described
in the CDISC standard for the variables USUBJID, PARCAT, PARAM, AVISIT and AVAL.
}

\subsection{resp_dataset}{

It expects a dataset similar to
https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192

It expects, at least, the columns passed in the parameters,
\code{subj_col}, \code{resp_cat_col}, \code{resp_par_col}, \code{resp_vis_col} and \code{resp_val}.
The values of these variables are as described
in the CDISC standard for the variables USUBJID, PARCAT, PARAM, AVISIT and AVAL.
}

\subsection{group_dataset}{

It expects a dataset with an structure similar to
https://www.cdisc.org/kb/examples/adam-subject-level-analysis-adsl-dataset-80283806 , one record per subject

It expects to contain, at least, \code{subj_col} and \code{group_col}
}
}

\section{Internal checks}{

\subsection{Shiny validation errors:}{
\itemize{
\item After selection it checks that:
\itemize{
\item Combination subject category parameter visit are unique for the predictor and response datasets
\item If grouped, that subject_id is unique for the selected group
\item The final selection contains at least one row
\item The response value is binary. Empty values are not considered for this check.
}
}
}

\subsection{Warnings:}{
\itemize{
\item Subjects in the selection have an empty response value
}
}
}

\keyword{internal}
