% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter.R
\name{scatterplot_UI}
\alias{scatterplot_UI}
\title{Scatter plot UI function}
\usage{
scatterplot_UI(id, default_include_regression_info)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}
}
\description{
Scatter plot UI function
}
\keyword{developers}
