% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter_matrix.R
\name{scatterplotmatrix_UI}
\alias{scatterplotmatrix_UI}
\title{Scatter plot matrix UI function}
\usage{
scatterplotmatrix_UI(id)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}
}
\description{
Scatter plot matrix UI function
}
\keyword{developers}
