% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter.R
\name{sp_subset_data}
\alias{sp_subset_data}
\title{Subset datasets for scatterplot}
\usage{
sp_subset_data(
  x_cat,
  y_cat,
  cat_col,
  x_par,
  y_par,
  par_col,
  x_val_col,
  y_val_col,
  x_vis,
  y_vis,
  vis_col,
  group_vect,
  bm_ds,
  group_ds,
  subj_col
)
}
\arguments{
\item{x_cat, y_cat, x_par, y_par, x_val_col, y_val_col, x_vis, y_vis}{Are similar to those without prefix in \code{\link[=subset_bds_param]{subset_bds_param()}}.}

\item{group_vect}{\verb{[named(character(n))]}

Columns to be subset and renamed.}

\item{bm_ds, group_ds}{\code{data.frame}

data frames to be used as inputs in \link{subset_bds_param} and \link{subset_adsl}}

\item{subj_col, par_col, cat_col, vis_col}{\verb{[character(1)]}

Column for subject id, category, parameter and visit. All specified columns must be factors}
}
\value{
\verb{[data.frame()]}

The \verb{_group} columns depend on the names in \code{group_vect}\tabular{lllll}{
   \code{subject_id} \tab \code{x_value} \tab \code{y_value} \tab \code{main_group} \tab \code{sub_group} \cr
   xx \tab xx \tab xx \tab xx \tab xx \cr
}
}
\description{
Prepares the basic input for the rest of the scatterplot functions.
\itemize{
\item \code{bm_dataset} is subset according to category, parameter and visit selection for x and y val
\item \code{group_dataset} is subset according to group_selection
\item both are joined using \code{subject_col} as a common key
\item it uses a left join
}

It is based on \code{\link[=subset_bds_param]{subset_bds_param()}} and \code{\link[=subset_adsl]{subset_adsl()}} with additional error checking.
}
\details{
\itemize{
\item factors from \code{bm_ds} are releveled so all extra levels not present after subsetting are dropped and are sorted
according to \code{par} and \code{cat}. Unless parameters are renamed in \code{\link[=subset_bds_param]{subset_bds_param()}} then no releveling occurs.
\item \code{group_vect} names are a subset of main_group and sub_group or empty,
otherwise a regular error is produced.
\item \code{label} attributes from \code{group_ds} and \code{bm_ds} are retained when available
}
\subsection{Shiny validation errors:}{
\itemize{
\item The x or y fragments from bm contains more than row per subject, category, parameter and visit combination
\item The fragment from group contains more than row per subject
\item If \code{bm_ds} and \code{grp_ds} share column names, apart from \code{subj_col}, after internal renaming has occured
\item If the returned dataset has 0 rows
}
}
}
\keyword{internal}
