% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{str_to_hash}
\alias{str_to_hash}
\title{Vectorized hash}
\usage{
str_to_hash(object)
}
\arguments{
\item{object}{object to be hashed}
}
\description{
Uses digest with a fixed murmur32 algorithm and no serialization
}
\keyword{internal}
