% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_logic.R
\name{strip_data_pronoun}
\alias{strip_data_pronoun}
\title{Strip ".data" pronoun from a string}
\usage{
strip_data_pronoun(x)
}
\arguments{
\item{x}{A character string to be processed.}
}
\value{
A character string with the ".data" pronoun removed.
}
\description{
This function removes the \code{.data} pronoun and the \verb{[[]]} accesor from a string using a regular expression pattern.
It works also when the .data is surrounded by \code{c()}. If there is no match the string is returned unmodified.
}
\keyword{internal}
