% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_logic.R
\name{subset_adsl}
\alias{subset_adsl}
\title{Subsets a group dataset, usually adsl, for a group_column_selection}
\usage{
subset_adsl(ds, group_vect, subj_col)
}
\arguments{
\item{ds}{\verb{[data.frame()]}

Data frame to be subset, see section \emph{Input dataframes}}

\item{group_vect}{\verb{[named(character(n))]}

Columns to be subset and renamed.}

\item{subj_col}{\verb{[character(1)]}

Column for subject id. \code{subj_col} must be a factor}
}
\value{
\code{data.frame()}\tabular{llll}{
   \code{subject_id} \tab \code{names(group_vec)[[1]]} \tab \code{names(group_vec)[[2]]} \tab ... \cr
   xx \tab xx \tab xx \tab xx \cr
}

\itemize{
\item When present \code{label} attribute is retained.
}
}
\description{
Subsets a group dataset, usually adsl, for a group_column_selection
}
\section{Input dataframe}{


It expects a dataset with an structure similar to
https://www.cdisc.org/kb/examples/adam-subject-level-analysis-adsl-dataset-80283806 , one record per subject

It expects to contain, at least, \code{subj_col} and all entries in \code{group_vect} as columns
}

\keyword{internal}
