% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_logic.R
\name{test_disjunct_cols}
\alias{test_disjunct_cols}
\alias{need_disjunct_cols}
\title{Test has repeated cols}
\usage{
test_disjunct_cols(ds1, ds2, ignore = NULL)

need_disjunct_cols(..., msg)
}
\arguments{
\item{ds1, ds2}{datframes to test}

\item{ignore}{columns to ignore while testing}

\item{...}{Argument forwarded to test_disjunct_cols}

\item{msg}{Validation message}
}
\description{
Test has repeated cols
}
\keyword{internal}
