% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_logic.R, R/mod_roc.R
\name{test_one_cat_per_par}
\alias{test_one_cat_per_par}
\alias{need_one_cat_per_var}
\title{Test if a parameter only appears under one category}
\usage{
test_one_cat_per_par(ds, cat_col, par_col)

need_one_cat_per_var(..., msg)

test_one_cat_per_par(ds, cat_col, par_col)
}
\arguments{
\item{ds}{\verb{[data.frame()]}
The dataset to be tested}

\item{cat_col}{\verb{[character(1)]}
Name of the column containing the category name}

\item{par_col}{\verb{[character(1)]}
Name of the column containing the parameter name}

\item{...}{arguments passed to test_one_cat_per_par}

\item{msg}{\verb{[character(1)]}

Validation message to be diplayed}
}
\value{
\verb{[logical(1)]}
\code{TRUE} if a parameter appears under a single category, \code{FALSE} otherwise

\verb{[logical(1)]}
\code{TRUE} if a parameter appears under a single category, \code{FALSE} otherwise
}
\description{
It contains at least the subject id. It includes a need to use inside a validate body

It contains at least the subject id
}
\keyword{helper}
\keyword{internal}
