% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_logic.R
\name{test_one_row_per_sbj}
\alias{test_one_row_per_sbj}
\alias{need_one_row_per_sbj}
\title{Test if a group has a single row for the combinations of a set of grouping variables}
\usage{
test_one_row_per_sbj(ds, subj_col, ...)

need_one_row_per_sbj(..., msg)
}
\arguments{
\item{ds}{\verb{[data.frame()]}
The dataset to be tested}

\item{subj_col}{\verb{[character(1)]}
Name of the column containing the subject_id}

\item{...}{\verb{[character(1)]}
Other column names for grouping}

\item{msg}{\verb{[character(1)]}
Validation message}
}
\value{
\verb{[logical(1)]}
\code{TRUE} if a single row is found, \code{FALSE} otherwise
}
\description{
It contains at least the subject id. It includes a need version to include in a validate body.
}
\keyword{internal}
