% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transformation.R
\name{transformation}
\alias{transformation}
\alias{tr_identity}
\alias{tr_z_score}
\alias{tr_gini}
\alias{tr_trunc_z_score}
\alias{tr_trunc_z_score_3_3}
\alias{tr_min_max}
\alias{tr_percentize}
\title{Transformation functions}
\usage{
tr_identity(x)

tr_z_score(x)

tr_gini(x)

tr_trunc_z_score(x, trunc_min = -3, trunc_max = 3)

tr_trunc_z_score_3_3(x)

tr_min_max(x)

tr_percentize(x)
}
\arguments{
\item{x}{a vector/list}

\item{trunc_min}{minimum value}

\item{trunc_max}{maximum value}
}
\value{
a vector/list with the transformed values

a vector/list with the transformed values

a vector/list with the transformed values truncated at the specified cuts

a vector/list with the transformed values truncated at the (-3, 3) cuts

a vector/list with the transformed values

a vector/list with the transformed values
}
\description{
Transformation functions
}
\section{Functions}{
\itemize{
\item \code{tr_identity()}: Identity transformation

\item \code{tr_z_score()}: Z score transformation

\item \code{tr_gini()}: Gini's mean difference

\item \code{tr_trunc_z_score()}: Truncated Z score transformation

\item \code{tr_trunc_z_score_3_3()}: Truncated Z score transformation in the (-3, 3) range

\item \code{tr_min_max()}: Min Max transformation

\item \code{tr_percentize()}: Percentize transformation

}}
