% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transformation.R
\name{transformation_utils}
\alias{transformation_utils}
\alias{get_tr_fun}
\alias{get_tr_apply}
\title{Transformation auxiliary functions}
\usage{
get_tr_fun(s)

get_tr_apply(tr)
}
\arguments{
\item{s}{the selected transformation}

\item{tr}{a function that has as first parameter the values to transform.}
}
\value{
a function from \link{get_tr_apply}

a function with the following interface \verb{function(df, group_col, val_col, ...)} where \code{...} is passed to as
extra parameters to tr
}
\description{
Transformation auxiliary functions
}
\section{Functions}{
\itemize{
\item \code{get_tr_fun()}: Transformation selecting function

\item \code{get_tr_apply()}: Transformation apply function

Get a function that applies a transformation to a given column grouped by another column

}}
\keyword{internal}
