% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_hmcat_server}
\alias{wfphm_hmcat_server}
\title{Waterfall plus heatmap categorical heatmap server function}
\usage{
wfphm_hmcat_server(
  id,
  dataset,
  subjid_var,
  sorted_x,
  cat_palette = list(),
  margin
)
}
\description{
Waterfall plus heatmap categorical heatmap server function
}
\keyword{developers}
