% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_hmcat_UI}
\alias{wfphm_hmcat_UI}
\title{Waterfall plus heatmap categorical heatmap UI function}
\usage{
wfphm_hmcat_UI(id)
}
\description{
Waterfall plus heatmap categorical heatmap UI function
}
\keyword{developers}
