% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_hmcont}
\alias{wfphm_hmcont}
\alias{wfphm_hmcont_UI}
\alias{wfphm_hmcont_server}
\title{Continuous heatmap component of WFPHM}
\usage{
wfphm_hmcont_UI(id)

wfphm_hmcont_server(id, dataset, subjid_var, sorted_x, margin)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{dataset}{\verb{[shiny::reactive(data.frame) | shinymeta::metaReactive(data.frame)]}

It expects the following format:
\itemize{
\item it contains, at least, the columns specified in the parameters: \code{subjid_var}
\item \code{subjid_var} columns is a factor
}}

\item{subjid_var}{\verb{[character(1)]}

column used as indicated in the details section}

\item{sorted_x}{\verb{[factor(*) | shiny::reactive(factor(*)) | shinymeta::metaReactive(factor(*))]}

indicates how the levels of \code{subjid_var} should be ordered in the X axis.}

\item{margin}{\verb{[numeric(4) | shiny::reactive(numeric(4)) | shinymeta::metaReactive(numeric(4))]}

margin to be used on each of the sides. It must contain four entries named \code{top}, \code{bottom}, \code{left} and \code{right}}
}
\value{
\subsection{UI}{

A heatmap plot
}

\subsection{Server}{

A list with one entry:
\itemize{
\item \code{margin}: similar to the \code{margin} parameter with the minimum margins for the current plot
}
}
}
\description{
Presents a heatmap plot for continuous variables
}
\section{Functions}{
\itemize{
\item \code{wfphm_hmcont_UI()}: UI

\item \code{wfphm_hmcont_server()}: server

}}
\section{Data subsetting:}{
\itemize{
\item Allows selecting several columns from all numerical columns, from now on \strong{numerical_selection}.
\itemize{
\item Menu labelled: Continuous heatmap
}
\item Subsets the dataset to the columns \strong{numerical_selection}, and \code{subj_var}.
\item Pivot the dataset to a longer format such that each row have:
\itemize{
\item column \code{x}: the value of \code{subjid_var}
\item column \code{y}: the value of a \strong{numerical_selection}
\item column \code{z}: the value of the \strong{numerical_selection} in \code{y} for the value in \code{x}
}
\item If more than one row have the same \code{x} and \code{y} an informative error indicating the plot cannot be created
is shown.
}
\subsection{X axis ordering}{
\itemize{
\item \code{x} levels are ordered according to \code{sorted_x}
}

Then a call to \link{heatmap_D3} is done with the following arguments:
\itemize{
\item \code{data} = \verb{subset dataset} (as described above)
\item \code{x_axis} = \code{NULL}
\item \code{y_axis} = \code{W}
\item \code{z_axis} = \code{E}
\item \code{margin} is the parameter passed to this same function
\item \code{palette} is hardcoded to \code{RColorBrewer::brewer.pal(n = 8, name = "Accent")} after 8 categories
colors are repeated
\item \code{msg_func} = NULL
\item \code{quiet} = TRUE
}
}
}

\keyword{internal}
