% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_hmcont_subset}
\alias{wfphm_hmcont_subset}
\title{Prepares the data for the continuous heatmap}
\usage{
wfphm_hmcont_subset(data, selection, subjid_var, sorted_x)
}
\arguments{
\item{data}{the dataframe to be subset, commonly a subject level dataset}

\item{selection}{the categorical columns to be selected}

\item{subjid_var}{the column that corresponde to the subject id}

\item{sorted_x}{the ordered subject ids}
}
\description{
Prepares the data for the continuous heatmap
}
\keyword{internal}
