% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_hmpar}
\alias{wfphm_hmpar}
\alias{wfphm_hmpar_UI}
\alias{wfphm_hmpar_server}
\title{Parameter heatmap component of WFPHM}
\usage{
wfphm_hmpar_UI(id, tr_choices)

wfphm_hmpar_server(
  id,
  dataset,
  cat_var,
  par_var,
  visit_var,
  anlfl_reactive = NULL,
  subjid_var,
  value_vars,
  sorted_x,
  tr_mapper,
  margin,
  show_x_ticks
)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{dataset}{\verb{[shiny::reactive(data.frame) | shinymeta::metaReactive(data.frame)]}

It expects the following format:
\itemize{
\item it contains, at least, the columns specified in the parameters: \code{cat_var}, \code{par_var}, \code{value_vars}, \code{visit_var}
and \code{subjid_var}
\item \code{cat_var}, \code{par_var}, \code{visit_var} and \code{subjid_var} columns are factors
\item \code{value_vars} must be numeric
}}

\item{cat_var, par_var, visit_var, subjid_var}{\verb{[character(1)]}}

\item{value_vars}{\verb{[character(1+)]}

possible colum values. If column is labelled, label will be displayed in the value menu}

\item{tr_mapper}{\verb{[function(1+)]}

named list containing a set of unary functions to transform the data. The name is the string shown in a selector and
the value is the function to be applied according to details section.}

\item{margin}{\verb{[numeric(4) | shiny::reactive(numeric(4)) | shinymeta::metaReactive(numeric(4))]}

margin to be used on each of the sides. It must contain four entries named \code{top}, \code{bottom}, \code{left} and \code{right}}

\item{show_x_ticks}{\verb{[logical(1)]}

show x ticks in the parameter heatmap}
}
\value{
\subsection{UI}{

A heatmap plot
}

\subsection{Server}{

A list with one entry:
\itemize{
\item \code{margin}: similar to the \code{margin} parameter with the minimum margins for the current plot
}
}
}
\description{
Parameter heatmap component of WFPHM
}
\details{
\subsection{Data subsetting:}{
\itemize{
\item Allows selecting several values from the \code{cat_var} column, from now on \strong{cat_selection} and several values
from the \code{par_var} column from the subset of rows where \code{par_cat} is equal to \strong{cat_selection}
from now on \strong{par_selection}.
\itemize{
\item Menu labelled: Category and Parameter
}
\item Allows selecting between the columns defined in \code{value_vars} from now on \strong{value_selection}.
\itemize{
\item Menu labelled: Value
}
\item Allows selecting a value from \code{visit_var} column, from now on \strong{visit_selection}.
\itemize{
\item Menu labelled: Visit
}
\item Subsets the dataset rows where:
\itemize{
\item \code{visit_var} equal to \strong{visit_selection}
\item \code{par_var} equal to \strong{par_selection}
}
\item Subsets the dataset to \code{par_var}, \code{subj_var} and the \strong{value_selection}.
\item If more than one row have the same combination \code{subj_var} and \code{par_var} an informative error indicating the plot
cannot be created is shown.
}

Then the dataset is prepared to be passed to \link{heatmap_D3}:
\itemize{
\item \code{subj_var} becomes \code{x} column
\item the label attribute of \code{x} column is \verb{Subject ID}
\item \code{par_var} becomes \code{y} column
\item \code{value_selection} becomes \code{z} column
\item \code{z} becomes \code{label} column
}
}

\subsection{Completing the dataset:}{
\itemize{
\item Subset dataset will be completed in the following way. All non-present combination of the original levels of \code{x}
and 'y' is are added with rows where:
\item \code{x} and \code{y} are equal to the missing combination
\item \code{z} is NA
}
}

\subsection{Data outliers:}{
\itemize{
\item Allows setting two limits upper and lower for \code{y} value, values above or below in the subsetted dataset will be
considered outliers. Rows considered outliers will have the column:
\itemize{
\item \code{label} replaced by "x"
\item \code{color} equal to "fill: white; font-weight: bold;"
}
\item Rows not considered outliers will have the column:
\itemize{
\item \code{color} equal to NA
}
}
}

\subsection{Data transformation:}{
\itemize{
\item Allows selecting between a set of functions as defined in \code{tr_mapper} from now on \strong{transformation_function}.
\itemize{
\item Menu labelled: Transformation
}
\item This transformation is applied to the subset dataset grouped by the \code{y} column. i.e. each row of the heatmap is
transformed independently.
\item The function must be unary, and the unique argument will be the numerical values of a given row of the hetamap.
}
}

\subsection{X axis ordering}{
\itemize{
\item \code{x} levels are ordered according to \code{sorted_x}
}

Then a call to \link{heatmap_D3} is done with the following arguments:
\itemize{
\item \code{data} = \verb{subset dataset} (as described above)
\item \code{x_axis} = \code{S}
\item \code{y_axis} = \code{W}
\item \code{z_axis} = \code{E}
\item \code{margin} is the parameter passed to this same function
\item \code{palette} is hardcoded to \code{RColorBrewer::brewer.pal(n = 8, name = "Set2")} after 8 categories colors are repeated
\item \code{msg_func} = NULL
\item \code{quiet} = TRUE
}
}
}
\section{Functions}{
\itemize{
\item \code{wfphm_hmpar_UI()}: UI

\item \code{wfphm_hmpar_server()}: server

}}
\keyword{internal}
