% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_server}
\alias{wfphm_server}
\title{Waterfall plus heatmap server function}
\usage{
wfphm_server(
  id,
  bm_dataset,
  group_dataset,
  cat_var = "PARCAT1",
  par_var = "PARAM",
  visit_var = "AVISIT",
  anlfl_vars = NULL,
  subjid_var = "USUBJID",
  value_vars = "AVAL",
  bar_group_palette = list(),
  cat_palette = list(),
  tr_mapper = tr_mapper_def(),
  show_x_ticks = TRUE
)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{bm_dataset}{\verb{[shiny::reactive(data.frame) | shinymeta::metaReactive(data.frame)]}}

\item{group_dataset}{\verb{[shiny::reactive(data.frame) | shinymeta::metaReactive(data.frame)]}}

\item{cat_var, par_var, visit_var, subjid_var}{\verb{[character(1)]}

columns used as indicated in each of the subplots}

\item{anlfl_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to analysis flags}

\item{value_vars}{\verb{[character(1+)]}

possible colum values. If column is labelled, label will be displayed in the value menu}

\item{bar_group_palette}{\verb{[list(palettes)]}

list of custom palettes to apply to bar_grouping. It receives the values used for grouping and must return a DaVinci
palette. Each palette is applied when the name of the entry in the list matches the name of the column used for
grouping}

\item{cat_palette}{\verb{[list(functions)]}

list of functions that receive the values of the variable and returns a vector with the colors for each of the values.
Each palette is applied when the name of the entry in the list matches the name of the selected categorical
variable}

\item{tr_mapper}{\verb{[function(1+)]}

named vector containing a set of transformation where the name is the string shown in the selector and the value is
function to be applied according to details section.}

\item{show_x_ticks}{\verb{[logical(1)]}

show x ticks in the parameter heatmap}
}
\description{
Waterfall plus heatmap server function
}
\keyword{developers}
