% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_UI}
\alias{wfphm_UI}
\title{Waterfall plus heatmap UI function}
\usage{
wfphm_UI(id, tr_choices = names(tr_mapper_def()))
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{tr_choices}{the names of the entries in tr_mapper}
}
\description{
Waterfall plus heatmap UI function
}
\keyword{developers}
