% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_wf}
\alias{wfphm_wf}
\alias{wfphm_wf_UI}
\title{Waterfall component of WFPHM}
\usage{
wfphm_wf(
  id,
  dataset,
  cat_var,
  par_var,
  visit_var,
  subjid_var,
  value_vars,
  bar_group_palette,
  margin
)

wfphm_wf_UI(id)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{dataset}{\verb{[shiny::reactive(data.frame) | shinymeta::metaReactive(data.frame)]}

It expects the following format:
\itemize{
\item it contains, at least, the columns specified in the parameters: \code{cat_var}, \code{par_var}, \code{value_vars}, \code{visit_var}
and \code{subjid_var}
\item \code{cat_var}, \code{par_var}, \code{visit_var} and \code{subjid_var} columns are factors
}}

\item{cat_var, par_var, visit_var, subjid_var}{\verb{[character(1)]}

columns used as indicated in the details section}

\item{value_vars}{\verb{[character(1+)]}

possible colum values. If column is labelled, label will be displayed in the value menu}

\item{bar_group_palette}{\verb{[list(palettes)]}

list of custom palettes to apply to bar_grouping. It receives the values used for grouping and must return a DaVinci
palette. Each palette is applied when the name of the entry in the list matches the name of the column used for
grouping}

\item{margin}{\verb{[numeric(4) | shiny::reactive(numeric(4)) | shinymeta::metaReactive(numeric(4))]}

margin to be used on each of the sides. It must contain four entries named \code{top}, \code{bottom}, \code{left} and \code{right}}
}
\value{
\subsection{UI}{

A bar plot
}

\subsection{Server}{

A list with two entries:
\itemize{
\item \code{margin}: similar to the \code{margin} parameter with the minimum margins for the current plot
\item \code{sorted_x}: From the subsetted data the \code{x} levels sorted from high to low by \code{y}
}
}
}
\description{
Waterfall component of WFPHM

Waterfall UI function
}
\details{
Data subsetting:
\itemize{
\item Allows selecting a column from all factor or character columns, from now on \strong{grouping_selection}.
\itemize{
\item Menu labelled: Grouping
}
\item Data selection for plotting can be done in two modes switched by:
\itemize{
\item Menu labelled: Display demographic baseline information
}
}
\subsection{Mode 1:}{
\itemize{
\item Allows selecting a column from the set of all numerical columns, from now on \strong{value_selection}.
\itemize{
\item Menu labelled: Value
}
\item Subsets the dataset to the columns \strong{grouping_selection}, \code{subj_var} and \strong{value_selection}.
\item Removes all repeated rows from the dataset
\item If more than one row have the same \code{subj_var} an informative error indicating the plot cannot be created
is shown.
}
}

\subsection{Mode 2:}{
\itemize{
\item Allows selecting a value from the \code{cat_var} column, from now on \strong{cat_selection} and a value from the \code{par_var}
column from the subset of rows where \code{par_cat} is equal to \strong{cat_selection} from now on \strong{par_selection}.
\itemize{
\item Menu labelled: Category and Parameter
}
\item Allows selecting between the columns defined in \code{val_var} from now on \strong{value_selection}.
\itemize{
\item Menu labelled: Value
}
\item Allows selecting a value from \code{visit_var} column, from now on \strong{visit_selection}.
\itemize{
\item Menu labelled: Visit
}
\item Subsets the dataset rows where:
\itemize{
\item \code{visit_var} equal to \strong{visit_selection}
\item \code{par_var} equal to \strong{par_selection}
}
\item Subsets the dataset to the \strong{grouping_selection}, the \code{subj_var} and the \strong{value_selection}.
\item If more than one row have the same \code{subj_var} an informative error indicating the plot cannot be created
is shown.
}

Then the dataset is prepared to be passed to \link{bar_D3}:
\itemize{
\item \code{subj_var} becomes \code{x} column
\item \code{val_selection} becomes \code{y} column
\item the label attribute of \code{y} column is either \code{value_selection} in \emph{Mode 1} or \code{par_selection} in \emph{Mode 2}
\item \code{grouping_selection} becomes \code{z} column
\item \code{grouping_selection} becomes \code{label} column
}
}

\subsection{Completing the dataset:}{
\itemize{
\item Subset dataset will be completed in the following way. If any level in \code{x} is not present in the subset
dataset, but it was present in the \code{subj_var} column in the original dataset, a row is added where \code{x} is equal to
the missing value \code{y} is NA and \code{z} is NA.
}
}

\subsection{Data outliers:}{
\itemize{
\item Allows setting two limits upper and lower, values above or below in the subsetted dataset will be considered
outliers. Rows considered outliers will have the column:
\itemize{
\item \code{label} replaced by "outlier"
\item \code{color} equal to "gray"
}
\item Rows not considered outliers will have the column:
\itemize{
\item \code{color} equal to NA
}
}
}

\subsection{X axis sorting}{
\itemize{
\item \code{x} levels are sorted from greater to lower value in \code{y}
}

Then a call to \link{bar_D3} is done with the following arguments:
\itemize{
\item \code{data} = \verb{subset dataset} (as described above)
\item \code{x_axis} = \code{NULL}
\item \code{y_axis} = \code{W}
\item \code{z_axis} = NULL
\item \code{margin} is the parameter passed to this same function
\item \code{palette} is hardcoded with 8 colors. After 8 categories colors are repeated
\item \code{msg_func} = NULL
\item \code{quiet} = TRUE
}
}
}
\section{Functions}{
\itemize{
\item \code{wfphm_wf_UI()}: UI

}}
\keyword{developers}
\keyword{internal}
