% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_wf_apply_outliers}
\alias{wfphm_wf_apply_outliers}
\title{Color bars and changes labels to outliers}
\usage{
wfphm_wf_apply_outliers(df, ll, ul)
}
\arguments{
\item{df}{the dataframe where we will apply the outliers}

\item{ll, ul}{the lower/upper value limit to consider an entry an outlier}
}
\description{
\code{label} is replaced by \code{outlier}
}
\details{
\code{color} is replaced by \code{gray}
}
\keyword{internal}
