% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_wf_rename_cols}
\alias{wfphm_wf_rename_cols}
\title{Renames wf subset columns}
\usage{
wfphm_wf_rename_cols(df)
}
\arguments{
\item{df}{the dataframe}
}
\description{
column names are replaced by x y z
}
\details{
y column label is set
}
\keyword{internal}
