% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_wf_server}
\alias{wfphm_wf_server}
\title{Waterfall server function}
\usage{
wfphm_wf_server(
  id,
  bm_dataset,
  group_dataset,
  cat_var,
  par_var,
  visit_var,
  subjid_var,
  value_vars,
  .default_group_palette = function(x) {
     pal_get_cat_palette(x,
    viridisLite::viridis(length(unique(x))))
 },
  bar_group_palette = list(),
  anlfl_reactive = NULL,
  margin
)
}
\description{
Waterfall server function
}
\keyword{developers}
