% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_wf_subset_data_cont}
\alias{wfphm_wf_subset_data_cont}
\title{Subsets a subject level dataset for the waterfall heatmap}
\usage{
wfphm_wf_subset_data_cont(val_col, color_col, group_ds, subj_col)
}
\arguments{
\item{val_col}{the column from the group_dataset to be used}

\item{color_col}{the column to be used for coloring/grouping}

\item{group_ds}{the subject level dataset}

\item{subj_col}{the subject id column}
}
\description{
Levels in subject column are ordered with respect to value column higher to lower
}
\keyword{internal}
