% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wfphm.R
\name{wfphm_wf_subset_data_par}
\alias{wfphm_wf_subset_data_par}
\title{Subset datasets for waterfall in waterfall plus heatmap}
\usage{
wfphm_wf_subset_data_par(
  cat,
  cat_col,
  par,
  par_col,
  val_col,
  vis,
  vis_col,
  color_col,
  bm_ds,
  group_ds,
  subj_col
)
}
\arguments{
\item{par, cat}{\verb{[character*ish*(n)]}

Values from \code{par_col} and \code{cat_col} to be subset}

\item{val_col}{\verb{[character*ish*(1)]}

Column containing values for the parameters}

\item{vis}{\verb{[character*ish*(1)]}

Values from \code{vis_col} to be subset}

\item{color_col}{the column used to color the bars}

\item{bm_ds, group_ds}{\code{data.frame}

data frames to be used as inputs in \link{subset_bds_param} and \link{subset_adsl}}

\item{subj_col, par_col, cat_col, vis_col}{\verb{[character(1)]}

Column for subject id, category, parameter and visit. All specified columns must be factors}
}
\value{
\verb{[data.frame()]}

The \verb{_group} columns depend on the names in \code{group_vect}\tabular{llll}{
   \code{x} \tab \code{y} \tab \code{z} \tab \code{color} \cr
   xx \tab xx \tab xx \tab xx \cr
}
}
\description{
Prepares the basic input for the rest of waterfall heatmap functions
\itemize{
\item \code{bm_dataset} is subset according to category, parameter and visit selection
\item \code{group_dataset} is subset according to group_selection
\item both are joined using \code{subject_col} as a common key
\item it uses a full join to include the subjects that has no parameter value for that parameter visit combination
}

It is based on \code{\link[=subset_bds_param]{subset_bds_param()}} and \code{\link[=subset_adsl]{subset_adsl()}} with additional error checking.
}
\details{
\itemize{
\item columns are renamed to \code{x}, \code{y}, \code{z} and \code{color}
\item \code{label} attributes from \code{group_ds} and \code{bm_ds} are retained when available
}
\subsection{Shiny validation errors:}{
\itemize{
\item The fragment from bm contains more than row per subject, category, parameter and visit combination
\item The fragment from group contains more than row per subject
\item If \code{bm_ds} and \code{grp_ds} share column names, apart from \code{subj_col}, after internal renaming has occured
\item If the returned dataset has 0 rows
}
}
}
\keyword{internal}
