% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{boxplot_server}
\alias{boxplot_server}
\title{Boxplot server function}
\usage{
boxplot_server(
  id,
  bm_dataset,
  group_dataset,
  dataset_name = shiny::reactive(character(0)),
  cat_var = "PARCAT",
  par_var = "PARAM",
  value_vars = "AVAL",
  x_axis_vars = NULL,
  anlfl_vars = NULL,
  subjid_var = "USUBJID",
  quantile_type = 7L,
  default_cat = NULL,
  default_par = NULL,
  default_x_axis_var = NULL,
  default_x_axis_vals = NULL,
  default_value = NULL,
  default_main_group = NULL,
  default_sub_group = NULL,
  default_page_group = NULL,
  on_sbj_click = function(x) {
 }
)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{bm_dataset, group_dataset}{\verb{[data.frame()]}

Dataframes as described in the \verb{Input dataframes} section}

\item{dataset_name}{\verb{[shiny::reactive(*)]}

a reactive indicating when the dataset has possibly changed its columns}

\item{cat_var, par_var}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category and parameter}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{x_axis_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to x-axis variables}

\item{anlfl_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to analysis flags}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to subject ID}

\item{quantile_type}{\verb{[integer(1)]}

Quantile algorithm type passed to \code{\link[stats]{quantile}}
(an integer between 1 and 9, default 7).}

\item{default_cat, default_par, default_x_axis_var, default_value, default_main_group, default_sub_group, default_page_group}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_x_axis_vals}{\verb{[character(n)|NULL]}

Default values for the x-axis variable}

\item{on_sbj_click}{\verb{[function()]}

Function to invoke when a subject is clicked in the single subject listing}
}
\description{
\subsection{Input dataframes:}{
\subsection{bm_dataset}{

It expects a dataset similar to
https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192 ,
1 record per subject per parameter per value of the selected x-axis variable.

It must contain, at least, the columns passed in the parameters, \code{subjid_var}, \code{cat_var}, \code{par_var},
\code{x_axis_vars} and \code{value_vars}. These correspond respectively to the subject identifier,
parameter category, parameter, x-axis grouping variables, and parameter values.
}

\subsection{group_dataset}{

It expects a dataset with an structure similar to
https://www.cdisc.org/kb/examples/adam-subject-level-analysis-adsl-dataset-80283806 ,
one record per subject.

It must contain, at least, the column passed in the parameter, \code{subjid_var}.
}

}
}
\keyword{developers}
