% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{boxplot_server}
\alias{boxplot_server}
\title{Boxplot server function}
\usage{
boxplot_server(
  id,
  bm_dataset,
  group_dataset,
  dataset_name = shiny::reactive(character(0)),
  cat_var = "PARCAT",
  par_var = "PARAM",
  value_vars = "AVAL",
  visit_var = "AVISIT",
  subjid_var = "USUBJID",
  default_cat = NULL,
  default_par = NULL,
  default_visit = NULL,
  default_value = NULL,
  default_main_group = NULL,
  default_sub_group = NULL,
  default_page_group = NULL,
  on_sbj_click = function(x) {
 }
)
}
\arguments{
\item{bm_dataset, group_dataset}{\verb{[data.frame()]}

Dataframes as described in the \verb{Input dataframes} section}

\item{dataset_name}{\verb{[shiny::reactive(*)]}

a reactive indicating when the dataset has possibly changed its columns}

\item{cat_var, par_var, visit_var, }{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to subject ID}

\item{default_cat, default_par, default_visit, default_value, default_main_group, default_sub_group, default_page_group}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{on_sbj_click}{\verb{[function()]}

Function to invoke when a subject is clicked in the single subject listing}
}
\description{
\subsection{Input dataframes:}{
\subsection{bm_dataset}{

It expects a dataset similar to
https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192 ,
1 record per subject per parameter per analysis visit.

It must contain, at least, the columns passed in the parameters, \code{subjid_var}, \code{cat_var}, \code{par_var},
\code{visit_var} and \code{value_vars}. The values of these variables are as described
in the CDISC standard for the variables USUBJID, PARCAT, PARAM, AVISIT and AVAL.
}

\subsection{group_dataset}{

It expects a dataset with an structure similar to
https://www.cdisc.org/kb/examples/adam-subject-level-analysis-adsl-dataset-80283806 ,
one record per subject.

It must contain, at least, the column passed in the parameter, \code{subjid_var}.
}

}
}
\keyword{developers}
