% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{bp_subset_data}
\alias{bp_subset_data}
\title{Subset datasets for boxplot}
\usage{
bp_subset_data(
  cat,
  cat_col,
  par,
  par_col,
  val_col,
  vis,
  vis_col,
  group_vect,
  bm_ds,
  group_ds,
  subj_col,
  anlfl_col = NULL
)
}
\arguments{
\item{par, cat}{\verb{[character*ish*(n)]}

Values from \code{par_col} and \code{cat_col} to be subset}

\item{val_col}{\verb{[character*ish*(1)]}

Column containing values for the parameters}

\item{vis}{\verb{[character*ish*(1)]}

Values from \code{vis_col} to be subset}

\item{group_vect}{\verb{[named(character(n))]}

Columns to be subset and renamed.}

\item{bm_ds, group_ds}{\code{data.frame}

data frames to be used as inputs in \link{subset_bds_param} and \link{subset_adsl}}

\item{subj_col, par_col, cat_col, vis_col, anlfl_col}{\verb{[character(1)]}

Column for subject id, category, parameter, visit and analysis flag. All
specified columns must be factors. Analysis flag is optional.}
}
\value{
\verb{[data.frame()]}

The \verb{_group} columns depend on the names in \code{group_vect}\tabular{llllll}{
   \code{subject_id} \tab \code{parameter} \tab \code{value} \tab \code{main_group} \tab \code{sub_group} \tab \code{page_group} \cr
   xx \tab xx \tab xx \tab xx \tab xx \tab xx \cr
}
}
\description{
Prepares the basic input for the rest of the boxplot functions.
\itemize{
\item \code{bm_dataset} is subset according to category, parameter and visit selection
\item \code{group_dataset} is subset according to group_selection
\item both are joined using \code{subject_col} as a common key
\item it uses a left join
}

It is based on \code{\link[=subset_bds_param]{subset_bds_param()}} and \code{\link[=subset_adsl]{subset_adsl()}} with additional error checking.
}
\details{
\itemize{
\item factors from \code{bm_ds} are releveled so all extra levels not present after subsetting are dropped and are sorted
according to \code{par} and \code{cat}. Unless parameters are renamed in \code{\link[=subset_bds_param]{subset_bds_param()}} then no releveling occurs.
\item \code{group_vect} names are a subset of main_group, sub_group and page_group or empty,
otherwise a regular error is produced.
\item \code{label} attributes from \code{group_ds} and \code{bm_ds} are retained when available
}
\subsection{Shiny validation errors:}{
\itemize{
\item The fragment from bm contains more than row per subject, category, parameter and visit combination
\item The fragment from group contains more than row per subject
\item If \code{bm_ds} and \code{grp_ds} share column names, apart from \code{subj_col}, after internal renaming has occured
\item If the returned dataset has 0 rows
}
}
}
\keyword{internal}
