% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_boxplot.R
\name{mock_app_boxplot_mm}
\alias{mock_app_boxplot_mm}
\title{Mock mm boxplot app}
\usage{
mock_app_boxplot_mm(update_query_string = TRUE, anlfl_flags = FALSE)
}
\arguments{
\item{update_query_string}{automatically update query string with app state}
}
\description{
Mock mm boxplot app
}
\keyword{mock}
