% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_corr_hm.R
\name{mod_corr_hm}
\alias{mod_corr_hm}
\title{Correlation Heatmap module}
\usage{
mod_corr_hm(
  module_id,
  bm_dataset_name,
  subjid_var = "USUBJID",
  cat_var = "PARCAT",
  par_var = "PARAM",
  visit_var = "AVISIT",
  anlfl_vars = NULL,
  value_vars = "AVAL",
  default_cat = NULL,
  default_par = NULL,
  default_visit = NULL,
  default_value = NULL
)
}
\arguments{
\item{module_id}{Shiny ID \verb{[character(1)]}

Module identifier}

\item{bm_dataset_name}{\verb{[character(1)]}

Biomarker dataset name}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to the subject ID}

\item{cat_var, par_var, visit_var}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{anlfl_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to analysis flags}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{default_cat}{Default selected categories}

\item{default_par}{Default selected parameters}

\item{default_visit}{Default selected visits}

\item{default_value}{\verb{[character(1)|NULL]}

Default values for the selectors}
}
\description{
Display a heatmap of correlation coefficients (Pearson, Spearman) along with confidence intervals
and p-values between dataset parameters over a single visit.
}
\keyword{main}
