% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_corr_hm.R
\name{mod_corr_hm}
\alias{mod_corr_hm}
\alias{corr_hm_UI}
\alias{corr_hm_server}
\title{Correlation Heatmap module}
\usage{
corr_hm_UI(
  id,
  default_cat = NULL,
  default_par = NULL,
  default_visit = NULL,
  default_corr_method = NULL
)

corr_hm_server(
  id,
  bm_dataset,
  subjid_var = "SUBJID",
  cat_var = "PARCAT",
  par_var = "PARAM",
  visit_var = "AVISIT",
  value_vars = c("AVAL", "PCHG"),
  default_value = NULL
)

mod_corr_hm(
  module_id,
  bm_dataset_name,
  subjid_var = "SUBJID",
  cat_var = "PARCAT",
  par_var = "PARAM",
  visit_var = "AVISIT",
  value_vars = c("AVAL", "PCHG"),
  default_cat = NULL,
  default_par = NULL,
  default_visit = NULL,
  default_value = NULL,
  bm_dataset_disp
)
}
\arguments{
\item{id}{\verb{[character(1)]}

Shiny ID}

\item{default_cat}{Default selected categories}

\item{default_par}{Default selected parameters}

\item{default_visit}{Default selected visits}

\item{default_corr_method}{Name of default correlation method}

\item{bm_dataset}{\verb{[data.frame()]}

An ADBM-like dataset similar in structure to the one in
\href{https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192}{this example},
with one record per subject per parameter per analysis visit.

It should have, at least, the columns specified by the parameters \code{subjid_var}, \code{cat_var},
\code{par_var}, \code{visit_var} and \code{value_vars}.
The semantics of these columns are as described in the CDISC standard for variables
USUBJID, PARCAT, PARAM, AVISIT and AVAL, respectively.}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to the subject ID}

\item{cat_var, par_var, visit_var}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{default_value}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{module_id}{Shiny ID \verb{[character(1)]}

Module identifier}

\item{bm_dataset_name}{\verb{[character(1)]}

Biomarker dataset name}

\item{bm_dataset_disp}{\verb{[mm_dispatcher(1)]}
module manager dispatchers passed as \code{bm_dataset} and \code{group_dataset} to \code{corr_hm_server}}
}
\description{
Display a heatmap of correlation coefficients (Pearson, Spearman) along with confidence intervals
and p-values between dataset parameters over a single visit.
}
\section{Functions}{
\itemize{
\item \code{corr_hm_UI()}: UI

\item \code{corr_hm_server()}: Server

}}
\keyword{main}
