% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineplot.R
\name{lineplot_UI}
\alias{lineplot_UI}
\alias{lineplot_server}
\alias{mod_lineplot}
\title{Line plot module}
\usage{
lineplot_UI(id)

lineplot_server(
  id,
  bm_dataset,
  group_dataset,
  dataset_name = shiny::reactive(character(0)),
  summary_functions = list(Mean = lp_mean_summary_functions, Median =
    lp_median_summary_functions),
  subjid_var = "SUBJID",
  cat_var = "PARCAT",
  par_var = "PARAM",
  visit_vars = c("AVISIT"),
  cdisc_visit_vars = character(0),
  value_vars = c("AVAL", "PCHG"),
  additional_listing_vars = character(0),
  ref_line_vars = character(0),
  on_sbj_click = NULL,
  default_centrality_function = NULL,
  default_dispersion_function = NULL,
  default_cat = NULL,
  default_par = NULL,
  default_val = NULL,
  default_visit_var = NULL,
  default_visit_val = NULL,
  default_main_group = NULL,
  default_sub_group = NULL,
  default_transparency = 1,
  default_y_axis_projection = "Linear"
)

mod_lineplot(
  module_id,
  bm_dataset_name,
  group_dataset_name,
  receiver_id = NULL,
  summary_functions = list(Mean = lp_mean_summary_functions, Median =
    lp_median_summary_functions),
  subjid_var = "SUBJID",
  cat_var = "PARCAT",
  par_var = "PARAM",
  visit_vars = c("AVISIT"),
  cdisc_visit_vars = character(0),
  value_vars = c("AVAL", "PCHG"),
  additional_listing_vars = character(0),
  ref_line_vars = character(0),
  default_centrality_function = NULL,
  default_dispersion_function = NULL,
  default_cat = NULL,
  default_par = NULL,
  default_val = NULL,
  default_visit_var = NULL,
  default_visit_val = NULL,
  default_main_group = NULL,
  default_sub_group = NULL,
  default_transparency = 1,
  default_y_axis_projection = "Linear"
)
}
\arguments{
\item{id}{Shiny ID \verb{[character(1)]}}

\item{bm_dataset}{\verb{[data.frame()]}

An ADBM-like dataset similar in structure to the one in
\href{https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192}{this example},
with one record per subject per parameter per analysis visit.

It should have, at least, the columns specified by the parameters \code{subjid_var}, \code{cat_var},
\code{par_var}, \code{visit_vars} and \code{value_vars}.
The semantics of these columns are as described in the CDISC standard for variables
USUBJID, PARCAT, PARAM, AVISIT and AVAL, respectively.

Optional columns specified by \code{ref_line_vars} should contain the same numeric value for all
records of the same parameter.}

\item{group_dataset}{\verb{[data.frame()]}

An ADSL-like dataset similar in structure to the one in
\href{https://www.cdisc.org/kb/examples/adam-subject-level-analysis-adsl-dataset-80283806}{this example},
with one record per subject.

It should contain, at least, the column specified by the parameter \code{subjid_var}.}

\item{dataset_name}{\verb{[shiny::reactive(*)]}

A reactive that indicates a possible change in the column structure of any of the two datasets}

\item{summary_functions}{\verb{[list()]}

Each element of this named list contains a summary function (e.g. a mean) and a collection of dispersion functions
(e.g. standard deviation) defining ranges around the values returned by the summary function.

The structure of each element is then a named list with the following elements:
\itemize{
\item \code{function}: Function that takes a numeric vector as its sole parameter and produces a scalar number.
\item \code{dispersion}: Named list of pairs functions that return the \emph{top} and \emph{bottom} dispersion ranges. They also take
a numeric vector as input and return a single numeric scalar
\item \code{y_prefix}: Prefix that will be prepended the Y axis label of the generated plot
}

For an example, see \code{dv.explorer.parameter::lp_mean_summary_functions}}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to the subject ID}

\item{cat_var, par_var, visit_vars}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{additional_listing_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that will be appended to the single-subject listing}

\item{ref_line_vars}{\verb{[character(n)]}

Columns for \code{bm_dataset} specifying reference values for parameters}

\item{on_sbj_click}{\verb{[function()]}

Function to invoke when a subject is clicked in the single-subject listing}

\item{default_cat, default_par, default_visit_var, default_visit_val, default_main_group}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_sub_group, default_val, default_centrality_function, default_dispersion_function}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_transparency}{\verb{[numeric(1)]}

Default values for the selectors}

\item{default_y_axis_projection}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{module_id}{Shiny ID \verb{[character(1)]}

Module identifier}

\item{bm_dataset_name, group_dataset_name}{\verb{[character(1)]}

Dataset names}

\item{receiver_id}{\verb{[character(1)]}

Name of the module receiving the selected subject ID in the single subject listing. The name must be present in
the module list or NULL.}
}
\description{
Display line plots of raw or summary data over time. Summaries include measures of central tendency
(mean, median) and optional deviation and confidence estimators.
}
\section{Functions}{
\itemize{
\item \code{lineplot_UI()}: UI

\item \code{lineplot_server()}: Server

}}
\keyword{main}
