% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roc.R
\name{mod_roc}
\alias{mod_roc}
\title{ROC module}
\usage{
mod_roc(
  module_id,
  pred_dataset_name,
  resp_dataset_name,
  group_dataset_name,
  pred_cat_var = "PARCAT",
  pred_par_var = "PARAM",
  pred_value_vars = "AVAL",
  pred_visit_var = "AVISIT",
  resp_cat_var = "PARCAT",
  resp_par_var = "PARAM",
  resp_value_vars = c("CHG1", "CHG2"),
  resp_visit_var = "AVISIT",
  subjid_var = "USUBJID",
  compute_roc_fn = compute_roc_data,
  compute_metric_fn = compute_metric_data
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

Module identificator}

\item{pred_dataset_name, resp_dataset_name, group_dataset_name}{Name of the dataset}

\item{pred_cat_var, pred_par_var, pred_visit_var, resp_cat_var, resp_par_var, resp_visit_var}{\verb{[character(1)]}

Columns from \code{pred_dataset}/\code{resp_dataset} that correspond to the parameter category, parameter and visit}

\item{pred_value_vars, resp_value_vars}{\verb{[character(n)]}

Columns from \code{pred_dataset},\code{resp_dataset} that correspond to values of the parameters}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to subject ID}

\item{compute_roc_fn, compute_metric_fn}{\verb{[function()]}

Functions used to compute the ROC and metric analysis, please view the corresponding vignette for more details.}
}
\description{
ROC module

Invoke ROC module
}
\keyword{main}
