% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter.R
\name{scatterplot_server}
\alias{scatterplot_server}
\title{Scatter plot server function}
\usage{
scatterplot_server(
  id,
  bm_dataset,
  group_dataset,
  dataset_name = shiny::reactive(character(0)),
  cat_var = "PARCAT",
  par_var = "PARAM",
  value_vars = "AVAL",
  visit_var = "AVISIT",
  anlfl_vars = NULL,
  subjid_var = "USUBJID",
  default_x_cat = NULL,
  default_x_par = NULL,
  default_x_value = NULL,
  default_x_visit = NULL,
  default_y_cat = NULL,
  default_y_par = NULL,
  default_y_value = NULL,
  default_y_visit = NULL,
  default_group = NULL,
  default_color = NULL,
  compute_lm_cor_fn = sp_compute_lm_cor_default
)
}
\arguments{
\item{bm_dataset, group_dataset}{\verb{[data.frame()]}

Dataframes as described in the \verb{Input dataframes} section}

\item{dataset_name}{\verb{[shiny::reactive(*)]}

a reactive indicating when the dataset has possibly changed its columns}

\item{cat_var, par_var, visit_var}{\verb{[character(1)]}

Columns from \code{bm_dataset} that correspond to the parameter category, parameter and visit}

\item{value_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to values of the parameters}

\item{anlfl_vars}{\verb{[character(n)]}

Columns from \code{bm_dataset} that correspond to analysis flags}

\item{subjid_var}{\verb{[character(1)]}

Column corresponding to subject ID}

\item{default_x_cat, default_x_par, default_x_visit, default_x_value, default_y_cat, default_y_par}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{default_y_visit, default_y_value, default_group, default_color}{\verb{[character(1)|NULL]}

Default values for the selectors}

\item{compute_lm_cor_fn}{\verb{[function()]}

Function used to compute the linear regression model and correlation statistics,
please view the corresponding vignette for more details.}
}
\description{
\subsection{Input dataframes:}{
\subsection{bm_dataset}{

It expects a dataset similar to
https://www.cdisc.org/kb/examples/adam-basic-data-structure-bds-using-paramcd-80288192 ,
1 record per subject per parameter per analysis visit

It expects, at least, the columns passed in the parameters, \code{subjid_var}, \code{cat_var}, \code{par_var},
\code{visit_var} and \code{value_var}. The values of these variables are as described
in the CDISC standard for the variables USUBJID, PARCAT, PARAM, AVISIT and AVAL.
}

\subsection{group_dataset}{

It expects a dataset with an structure similar to
https://www.cdisc.org/kb/examples/adam-subject-level-analysis-adsl-dataset-80283806 ,
one record per subject

It expects to contain, at least, \code{subjid_var}
}

}
}
\keyword{developers}
